queryset pattern for serializers, reorganize tests
