+++ iaso/api/profiles/serializers/update.py
@@ -4,7 +4,7 @@ from django.utils.translation import gettext_lazy as _
 from rest_framework import serializers
 
 from iaso.api.common import ModelSerializer
-from iaso.api.profiles.serializers.common import CountryAwarePhoneNumberField
+from iaso.api.common.serializer_fields import CountryAwarePhoneNumberField
 from iaso.models import OrgUnit, OrgUnitType, Profile, Project, TenantUser, UserRole
 
 
@@ -33,11 +33,16 @@ class ProfileUpdateSerializer(BaseProfileUpdateSerializer):
         allow_empty=True, allow_null=True, queryset=Project.objects.none(), required=False, many=True
     )
     user_permissions = serializers.SlugRelatedField(
-        slug_field="codename", many=True, queryset=Permission.objects.all(), required=False
+        slug_field="codename", many=True, queryset=Permission.objects.none(), required=False
     )
 
     user_roles = serializers.PrimaryKeyRelatedField(
-        allow_empty=True, allow_null=True, many=True, queryset=UserRole.objects.all(), required=False
+        allow_empty=True,
+        allow_null=True,
+        many=True,
+        queryset=UserRole.objects.none(),
+        required=False,
+        error_messages={"does_not_exist": _("One or more user roles do not belong to the provided account.")},
     )
 
     org_units = serializers.PrimaryKeyRelatedField(
@@ -78,45 +83,20 @@ class ProfileUpdateSerializer(BaseProfileUpdateSerializer):
         profile = self.instance
         if profile is not None:
             self.fields["projects"].child_relation.queryset = Project.objects.filter(account_id=profile.account_id)
+            self.fields["user_roles"].child_relation.queryset = UserRole.objects.filter(account=profile.account)
+            module_permission_codenames = {perm.codename for perm in profile.account.permissions_from_active_modules}
+            self.fields["user_permissions"].child_relation.queryset = Permission.objects.filter(
+                codename__in=list(module_permission_codenames)
+            )
 
     def validate_dhis2_id(self, value):
         return value or None
 
-    def validate_user_roles(self, data):
-        account = self._get_account()
-        if not data:
-            return data
-
-        data = set(data)
-
-        if UserRole.objects.filter(pk__in=[user_role.pk for user_role in data]).exclude(account=account).exists():
-            raise serializers.ValidationError(
-                {"user_roles": _("One or more user roles do not belong to the provided account.")}
-            )
-
-        return data
-
     def _get_account(self):
         request = self.context["request"]
-        current_profile = request.user.iaso_profile
-        return current_profile.account
-
-    def _validate_user_permissions(self, data):
-        account = self._get_account()
-        user_permissions = data.get("user_permissions")
-
-        if not account or not user_permissions:
-            return []
-
-        module_permission_codenames = {perm.codename for perm in account.permissions_from_active_modules}
-
-        valid_permissions = [perm for perm in user_permissions if perm.codename in module_permission_codenames]
-
-        return valid_permissions
-
-    def validate(self, data):
-        data["user_permissions"] = self._validate_user_permissions(data)
-        return data
+        user = getattr(request, "user", None)
+        current_profile = getattr(user, "iaso_profile", None)
+        return getattr(current_profile, "account", None)
 
     def validate_org_units(self, org_units):
         profile = self.instance
