--- iaso/tests/api/test_profiles.py
@@ -1,2596 +1,0 @@
-import typing
-
-import jsonschema
-
-from django.contrib.auth import get_user_model
-from django.contrib.auth.models import Group, Permission
-from django.core import mail
-from django.test import override_settings
-from django.urls import reverse
-from rest_framework import status
-
-from iaso import models as m
-from iaso.api.profiles.views import PK_ME
-from iaso.models import Profile
-from iaso.models.team import Team
-from iaso.modules import MODULES
-from iaso.permissions.core_permissions import (
-    CORE_FORMS_PERMISSION,
-    CORE_ORG_UNITS_READ_PERMISSION,
-    CORE_USERS_ADMIN_PERMISSION,
-    CORE_USERS_MANAGED_PERMISSION,
-)
-from iaso.test import APITestCase
-from iaso.utils.colors import DEFAULT_COLOR
-
-
-name_and_id_schema = {
-    "type": "object",
-    "properties": {"id": {"type": "number"}, "name": {"type": "string"}},
-    "required": ["name", "id"],
-}
-
-PROFILE_LOG_SCHEMA = {
-    "type": "object",
-    "properties": {
-        "id": {"type": "number"},
-        "content-type": {"type": "string"},
-        "object_id": {"type": "string"},
-        "source": {"type": "string"},
-        "created_at": {"type": "string"},
-        "user": {
-            "type": "object",
-            "properties": {
-                "id": {"type": "number"},
-                "first_name": {"type": ["string", "null"]},
-                "last_name": {"type": ["string", "null"]},
-                "user_name": {"type": "string"},
-                "email": {"type": ["string", "null"]},
-                "language": {"type": ["string", "null"]},
-                "phone_number": {"type": ["string", "null"]},
-                "country_code": {"type": ["string", "null"]},
-            },
-            "required": ["id", "user_name"],
-        },
-        "past_value": {
-            "type": "array",
-            "items": {
-                "type": "object",
-                "properties": {
-                    "pk": {"type": "number"},
-                    "fields": {
-                        "type": "object",
-                        "properties": {
-                            "user": {"type": "number"},
-                            "first_name": {"type": ["string", "null"]},
-                            "last_name": {"type": ["string", "null"]},
-                            "username": {"type": ["string", "null"]},
-                            "email": {"type": ["string", "null"]},
-                            "organization": {"type": ["string", "null"]},
-                            "user_permissions": {"type": "array", "items": {"type": "string"}},
-                            "deleted_at": {"type": ["string", "null"]},
-                            "account": {"type": "number"},
-                            "dhis2_id": {"type": ["string", "null"]},
-                            "language": {"type": ["string", "null"]},
-                            "home_page": {"type": ["string", "null"]},
-                            "projects": {"type": "array", "items": {"type": "number"}},
-                            "org_units": {"type": "array", "items": {"type": "number"}},
-                            "user_roles": {"type": "array", "items": {"type": "number"}},
-                            "phone_number": {"type": ["string", "null"]},
-                        },
-                        "required": ["user", "account"],
-                    },
-                },
-            },
-        },
-        "new_value": {
-            "type": "array",
-            "items": {
-                "type": "object",
-                "properties": {
-                    "pk": {"type": "number"},
-                    "fields": {
-                        "type": "object",
-                        "properties": {
-                            "user": {"type": "number"},
-                            "first_name": {"type": ["string", "null"]},
-                            "last_name": {"type": ["string", "null"]},
-                            "username": {"type": ["string", "null"]},
-                            "email": {"type": ["string", "null"]},
-                            "organization": {"type": ["string", "null"]},
-                            "user_permissions": {"type": "array", "items": {"type": "string"}},
-                            "password_updated": {"type": "boolean"},
-                            "deleted_at": {"type": ["string", "null"]},
-                            "account": {"type": "number"},
-                            "dhis2_id": {"type": ["string", "null"]},
-                            "language": {"type": ["string", "null"]},
-                            "home_page": {"type": ["string", "null"]},
-                            "projects": {"type": "array", "items": {"type": "number"}},
-                            "org_units": {"type": "array", "items": {"type": "number"}},
-                            "user_roles": {"type": "array", "items": {"type": "number"}},
-                            "phone_number": {"type": ["string", "null"]},
-                        },
-                        "required": ["user"],
-                    },
-                },
-            },
-        },
-    },
-    "required": ["id", "user", "content_type", "source", "object_id", "created_at", "past_value", "new_value"],
-}
-
-
-class ProfileAPITestCase(APITestCase):
-    @classmethod
-    def setUpTestData(cls):
-        cls.MODULES = [module.codename for module in MODULES]
-        cls.account = m.Account.objects.create(name="Global Health Initiative", modules=cls.MODULES)
-        cls.another_account = m.Account.objects.create(name="Another account")
-
-        # TODO : make the org unit creations shorter and reusable
-        cls.project = m.Project.objects.create(
-            name="Hydroponic gardens",
-            app_id="stars.empire.agriculture.hydroponics",
-            account=cls.account,
-        )
-        datasource = m.DataSource.objects.create(name="Evil Empire")
-        datasource.projects.add(cls.project)
-        cls.datasource = datasource
-        cls.sub_unit_type = m.OrgUnitType.objects.create(name="Jedi Squad", short_name="Jds")
-        cls.parent_org_unit_type = m.OrgUnitType.objects.create(name="Jedi Council", short_name="Cnc")
-        cls.parent_org_unit_type.sub_unit_types.add(cls.sub_unit_type)
-
-        cls.mock_multipolygon = None
-        cls.mock_point = None
-
-        cls.org_unit_group = m.Group.objects.create(name="Elite councils")
-        cls.datasource = datasource
-        source_version_1 = m.SourceVersion.objects.create(data_source=datasource, number=1)
-        cls.account.default_version = source_version_1
-        cls.account.save()
-        cls.org_unit_from_sub_type = m.OrgUnit.objects.create(
-            org_unit_type=cls.sub_unit_type,
-            version=source_version_1,
-            name="Jedi Squad 1",
-            geom=cls.mock_multipolygon,
-            simplified_geom=cls.mock_multipolygon,
-            catchment=cls.mock_multipolygon,
-            location=cls.mock_point,
-            validation_status=m.OrgUnit.VALIDATION_VALID,
-            source_ref=None,
-        )
-        cls.org_unit_type = m.OrgUnitType.objects.create(name="Org unit type")
-        cls.another_org_unit = m.OrgUnit.objects.create(
-            org_unit_type=cls.org_unit_type,
-            name="Hôpital Général",
-        )
-        cls.org_unit_from_parent_type = m.OrgUnit.objects.create(
-            org_unit_type=cls.parent_org_unit_type,
-            version=source_version_1,
-            name="Corruscant Jedi Council",
-            geom=cls.mock_multipolygon,
-            simplified_geom=cls.mock_multipolygon,
-            catchment=cls.mock_multipolygon,
-            location=cls.mock_point,
-            validation_status=m.OrgUnit.VALIDATION_VALID,
-            source_ref="FooBarB4z00",
-        )
-        cls.org_unit_from_parent_type.groups.set([cls.org_unit_group])
-
-        cls.child_org_unit = m.OrgUnit.objects.create(
-            org_unit_type=cls.parent_org_unit_type,
-            version=source_version_1,
-            name="Corruscant Jedi Council",
-            geom=cls.mock_multipolygon,
-            simplified_geom=cls.mock_multipolygon,
-            catchment=cls.mock_multipolygon,
-            location=cls.mock_point,
-            validation_status=m.OrgUnit.VALIDATION_VALID,
-            source_ref="PvtAI4RUMkr",
-            parent=cls.org_unit_from_parent_type,
-        )
-
-        cls.permission = Permission.objects.create(
-            name="iaso permission", content_type_id=1, codename="iaso_permission"
-        )
-        cls.group = Group.objects.create(name="user role")
-        cls.group.permissions.add(cls.permission)
-        cls.user_role = m.UserRole.objects.create(group=cls.group, account=cls.account)
-
-        cls.group_another_account = Group.objects.create(name="user role with another account")
-        cls.group_another_account.permissions.add(cls.permission)
-        cls.user_role_another_account = m.UserRole.objects.create(
-            group=cls.group_another_account, account=cls.another_account
-        )
-
-        # Users.
-        cls.jane = cls.create_user_with_profile(
-            first_name="Jane",
-            last_name="Doe",
-            username="janedoe",
-            account=cls.account,
-            permissions=[CORE_FORMS_PERMISSION],
-        )
-        cls.john = cls.create_user_with_profile(username="johndoe", account=cls.account, is_superuser=True)
-        cls.jim = cls.create_user_with_profile(
-            username="jim", account=cls.account, permissions=[CORE_FORMS_PERMISSION, CORE_USERS_ADMIN_PERMISSION]
-        )
-        cls.jam = cls.create_user_with_profile(
-            username="jam",
-            account=cls.account,
-            permissions=[CORE_USERS_MANAGED_PERMISSION],
-            language="en",
-        )
-        cls.jom = cls.create_user_with_profile(username="jom", account=cls.account, permissions=[], language="fr")
-        cls.jum = cls.create_user_with_profile(
-            username="jum", account=cls.account, permissions=[], projects=[cls.project]
-        )
-        cls.user_managed_geo_limit = cls.create_user_with_profile(
-            username="managedGeoLimit",
-            account=cls.account,
-            permissions=[CORE_USERS_MANAGED_PERMISSION],
-            org_units=[cls.org_unit_from_parent_type],
-        )
-        cls.team1 = Team.objects.create(project=cls.project, name="team1", manager=cls.jane)
-        cls.team1.users.add(cls.jane)
-        cls.team2 = Team.objects.create(project=cls.project, name="team2", manager=cls.jim)
-        cls.team2.users.add(cls.jim)
-        cls.user_managed_geo_limit.iaso_profile.user_roles.set([cls.user_role, cls.user_role_another_account])
-
-        cls.user_role_name = cls.user_role.group.name.removeprefix(
-            f"{cls.user_managed_geo_limit.iaso_profile.account.pk}_"
-        )
-        cls.user_role_another_account_name = cls.user_role_another_account.group.name.removeprefix(
-            f"{cls.user_managed_geo_limit.iaso_profile.account.pk}_"
-        )
-
-    def assertValidProfileListData(self, list_data: typing.Mapping, expected_length: int, paginated: bool = False):
-        self.assertValidListData(
-            list_data=list_data,
-            expected_length=expected_length,
-            results_key="results",
-            paginated=paginated,
-        )
-
-        for profile_data in list_data["results"]:
-            self.assertValidProfileData(profile_data)
-
-    def assertValidProfileData(self, project_data: typing.Mapping):
-        self.assertHasField(project_data, "id", int)
-        self.assertHasField(project_data, "firstName", str)
-        self.assertHasField(project_data, "lastName", str)
-        self.assertHasField(project_data, "email", str)
-        self.assertHasField(project_data, "color", str)
-
-    def get_new_user_data(self):
-        user_name = "audit_user"
-        pwd = "admin1234lol"
-        first_name = "audit_user_first_name"
-        last_name = "audit_user_last_name"
-        email = "audit@test.com"
-        organization = "Bluesquare"
-        org_units = [f"{self.org_unit_from_parent_type.id}"]
-        language = "fr"
-        home_page = "/forms/list"
-        dhis2_id = "666666666666"
-        user_roles = [self.user_role.id]
-        user_roles_permissions = [self.user_role.id]
-        user_permissions = [CORE_ORG_UNITS_READ_PERMISSION.codename]
-        send_email_invitation = False
-        projects = [self.project.id]
-        phone_number = "+32475888888"
-        country_code = "be"
-        data = {
-            "userName": user_name,
-            "password": pwd,
-            "firstName": first_name,
-            "lastName": last_name,
-            "sendEmailInvitation": send_email_invitation,
-            "email": email,
-            "organization": organization,
-            "language": language,
-            "homePage": home_page,
-            "dhis2Id": dhis2_id,
-            "permissions": [],  # This looks legacy from an older version of the API
-            "userPermissions": user_permissions,
-            "projects": projects,
-            "phoneNumber": phone_number,
-            "countryCode": country_code,
-            "userRoles": user_roles,
-            "userRolesPermissions": user_roles_permissions,
-            "orgUnits": org_units,
-        }
-        return data
-
-    def test_retrieve_profile_me_without_auth(self):
-        """GET /profiles/me/ without auth should result in a 401"""
-        response = self.client.get(reverse("profiles-detail", kwargs={"pk": "me", "version": "v2"}))
-        self.assertJSONResponse(response, 401)
-
-    def test_retrieve_me_is_compatible_for_mobile(self):
-        user = self.create_user_with_profile(
-            first_name="Jane",
-            last_name="Doe",
-            username="janedoe2",
-            account=self.account,
-            permissions=[CORE_FORMS_PERMISSION],
-        )
-        project_1 = m.Project.objects.create(name="Project 1", app_id="project.1", account=self.account)
-        user.iaso_profile.phone_number = "+32477123456"
-        user.iaso_profile.country_code = "be"
-        user.iaso_profile.projects.set([project_1])
-        user.iaso_profile.save()
-
-        self.client.force_authenticate(user)
-        response = self.client.get(reverse("profiles-detail", kwargs={"pk": "me"}))
-        response_data = self.assertJSONResponse(response, 200)
-        for k in ["id", "first_name", "last_name", "user_name", "email", "phone_number", "organization", "projects"]:
-            self.assertIn(k, response_data)
-
-        for k in ["id", "name", "app_id"]:
-            self.assertIn(k, response_data["projects"][0])
-
-    def test_retrieve_profile_me_ok(self):
-        """GET /profiles/me/ with auth"""
-
-        self.client.force_authenticate(self.jane)
-        response = self.client.get(reverse("profiles-detail", kwargs={"pk": "me", "version": "v2"}))
-        response_data = self.assertJSONResponse(response, 200)
-
-        self.assertValidProfileData(response_data)
-        self.assertEqual(response_data["userName"], "janedoe")
-        self.assertHasField(response_data, "account", dict)
-        self.assertHasField(response_data, "permissions", list)
-        self.assertHasField(response_data, "isSuperuser", bool)
-        self.assertHasField(response_data, "orgUnits", list)
-        self.assertEqual(response_data["color"], DEFAULT_COLOR)
-
-    def test_retrieve_profile_me_no_profile(self):
-        """GET /profiles/me/ with auth, but without profile
-        The goal is to know that this call doesn't result in a 500 error
-        """
-        username = "I don't have a profile, i'm sad :("
-        user_without_profile = get_user_model().objects.create(username=username)
-        self.client.force_authenticate(user_without_profile)
-        response = self.client.get(reverse("profiles-detail", kwargs={"pk": "me", "version": "v2"}))
-        response_data = self.assertJSONResponse(response, status.HTTP_200_OK)
-
-        self.assertEqual(response_data["userName"], username)
-        self.assertEqual(response_data["firstName"], "")
-        self.assertEqual(response_data["lastName"], "")
-        self.assertEqual(response_data["userId"], user_without_profile.id)
-        self.assertEqual(response_data["email"], "")
-        self.assertEqual(response_data["projects"], [])
-        self.assertFalse(response_data["isStaff"])
-        self.assertFalse(response_data["isSuperuser"])
-        self.assertIsNone(response_data["account"])
-
-    def test_retrieve_profile_me_superuser_ok(self):
-        """GET /profiles/me/ with auth (superuser)"""
-
-        self.client.force_authenticate(self.john)
-        response = self.client.get(reverse("profiles-detail", kwargs={"pk": "me", "version": "v2"}))
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileData(response_data)
-        self.assertEqual(response_data["userName"], "johndoe")
-
-    # todo : write tests for retrieve , but not for "me"
-
-    def test_can_delete_dhis2_id(self):
-        self.client.force_authenticate(self.john)
-        jim = Profile.objects.get(user=self.jim)
-        jim.dhis2_id = "random_dhis2_id"
-        jim.save()
-
-        data = {"dhis2_id": ""}
-
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": jim.id, "version": "v2"}), data=data, format="json"
-        )
-
-        self.assertJSONResponse(response, 200)
-
-    def test_profile_list_no_auth(self):
-        """GET /profiles/ without auth -> 401"""
-
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}))
-        self.assertJSONResponse(response, 401)
-
-    def test_profile_list_number_queries_without_read_only_permissions(self):
-        """GET /profiles/ with auth (user has read only permissions)"""
-        self.client.force_authenticate(self.jane)
-        with self.assertNumQueries(13):
-            response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}))
-        self.assertJSONResponse(response, 200)
-
-    def test_profile_retrieve_read_only_permissions(self):
-        """GET /profiles/ with auth (user has read only permissions)"""
-
-        self.client.force_authenticate(self.jane)
-        response = self.client.get(
-            reverse("profiles-detail", kwargs={"pk": self.jane.iaso_profile.id, "version": "v2"})
-        )
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileData(response_data)
-        self.assertEqual(response_data["userName"], "janedoe")
-
-    def test_profile_patchread_only_permissions(self):
-        """GET /profiles/ with auth (user has read only permissions)"""
-
-        self.client.force_authenticate(self.jane)
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": self.jane.iaso_profile.id, "version": "v2"})
-        )
-        self.assertJSONResponse(response, 403)
-
-    def test_profile_list_ok(self):
-        """GET /profiles/ with auth"""
-        self.client.force_authenticate(self.jane)
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}))
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 7)
-
-    def test_profile_list_export_as_csv(self):
-        self.maxDiff = None
-        self.john.iaso_profile.org_units.set([self.org_unit_from_sub_type, self.org_unit_from_parent_type])
-        self.jum.iaso_profile.editable_org_unit_types.set([self.sub_unit_type])
-
-        self.client.force_authenticate(self.jane)
-        response = self.client.get(reverse("profiles-export-csv", kwargs={"version": "v2"}))
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response["Content-Type"], "text/csv")
-
-        response_csv = response.getvalue().decode("utf-8")
-
-        expected_csv = "".join(
-            [
-                "user_profile_id,"
-                "username,"
-                "password,"
-                "email,"
-                "first_name,"
-                "last_name,"
-                "orgunit,"
-                "orgunit__source_ref,"
-                "profile_language,"
-                "dhis2_id,"
-                "organization,"
-                "permissions,"
-                "user_roles,"
-                "projects,"
-                "teams,"
-                "phone_number,"
-                "editable_org_unit_types\r\n"
-            ]
-        )
-
-        expected_csv += f"{self.jane.iaso_profile.id},janedoe,,,Jane,Doe,,,,,,iaso_forms,,,{self.team1.name},,\r\n"
-        expected_csv += f'{self.john.iaso_profile.id},johndoe,,,,,"{self.org_unit_from_sub_type.pk},{self.org_unit_from_parent_type.pk}",{self.org_unit_from_parent_type.source_ref},,,,,,,,,\r\n'
-        expected_csv += f'{self.jim.iaso_profile.id},jim,,,,,,,,,,"{CORE_FORMS_PERMISSION.codename},{CORE_USERS_ADMIN_PERMISSION.codename}",,,{self.team2.name},,\r\n'
-        expected_csv += f"{self.jam.iaso_profile.id},jam,,,,,,,en,,,{CORE_USERS_MANAGED_PERMISSION.codename},,,,,\r\n"
-        expected_csv += f"{self.jom.iaso_profile.id},jom,,,,,,,fr,,,,,,,,\r\n"
-        expected_csv += f"{self.jum.iaso_profile.id},jum,,,,,,,,,,,,{self.project.name},,,{self.sub_unit_type.pk}\r\n"
-        expected_csv += f'{self.user_managed_geo_limit.iaso_profile.id},managedGeoLimit,,,,,{self.org_unit_from_parent_type.id},{self.org_unit_from_parent_type.source_ref},,,,{CORE_USERS_MANAGED_PERMISSION.codename},"{self.user_role_name},{self.user_role_another_account_name}",,,,\r\n'
-
-        self.assertEqual(response_csv, expected_csv)
-
-    def test_profile_list_export_as_xlsx(self):
-        self.maxDiff = None
-        self.john.iaso_profile.org_units.set([self.org_unit_from_sub_type, self.org_unit_from_parent_type])
-        self.jum.iaso_profile.editable_org_unit_types.set([self.sub_unit_type])
-
-        self.client.force_authenticate(self.jane)
-        response = self.client.get(reverse("profiles-export-xlsx", kwargs={"version": "v2"}))
-        excel_columns, excel_data = self.assertXlsxFileResponse(response)
-
-        self.assertEqual(
-            excel_columns,
-            [
-                "user_profile_id",
-                "username",
-                "password",
-                "email",
-                "first_name",
-                "last_name",
-                "orgunit",
-                "orgunit__source_ref",
-                "profile_language",
-                "dhis2_id",
-                "organization",
-                "permissions",
-                "user_roles",
-                "projects",
-                "teams",
-                "phone_number",
-                "editable_org_unit_types",
-            ],
-        )
-
-        self.assertDictEqual(
-            excel_data,
-            {
-                "user_profile_id": {
-                    0: self.jane.iaso_profile.id,
-                    1: self.john.iaso_profile.id,
-                    2: self.jim.iaso_profile.id,
-                    3: self.jam.iaso_profile.id,
-                    4: self.jom.iaso_profile.id,
-                    5: self.jum.iaso_profile.id,
-                    6: self.user_managed_geo_limit.iaso_profile.id,
-                },
-                "username": {0: "janedoe", 1: "johndoe", 2: "jim", 3: "jam", 4: "jom", 5: "jum", 6: "managedGeoLimit"},
-                "password": {0: None, 1: None, 2: None, 3: None, 4: None, 5: None, 6: None},
-                "email": {0: None, 1: None, 2: None, 3: None, 4: None, 5: None, 6: None},
-                "first_name": {0: "Jane", 1: None, 2: None, 3: None, 4: None, 5: None, 6: None},
-                "last_name": {0: "Doe", 1: None, 2: None, 3: None, 4: None, 5: None, 6: None},
-                "orgunit": {
-                    0: None,
-                    1: f"{self.org_unit_from_sub_type.id},{self.org_unit_from_parent_type.id}",
-                    2: None,
-                    3: None,
-                    4: None,
-                    5: None,
-                    6: f"{self.org_unit_from_parent_type.id}",
-                },
-                "orgunit__source_ref": {
-                    0: None,
-                    1: self.org_unit_from_parent_type.source_ref,
-                    2: None,
-                    3: None,
-                    4: None,
-                    5: None,
-                    6: self.org_unit_from_parent_type.source_ref,
-                },
-                "profile_language": {0: None, 1: None, 2: None, 3: "en", 4: "fr", 5: None, 6: None},
-                "dhis2_id": {0: None, 1: None, 2: None, 3: None, 4: None, 5: None, 6: None},
-                "organization": {0: None, 1: None, 2: None, 3: None, 4: None, 5: None, 6: None},
-                "permissions": {
-                    0: CORE_FORMS_PERMISSION.codename,
-                    1: None,
-                    2: f"{CORE_FORMS_PERMISSION.codename},{CORE_USERS_ADMIN_PERMISSION.codename}",
-                    3: CORE_USERS_MANAGED_PERMISSION.codename,
-                    4: None,
-                    5: None,
-                    6: CORE_USERS_MANAGED_PERMISSION.codename,
-                },
-                "user_roles": {
-                    0: None,
-                    1: None,
-                    2: None,
-                    3: None,
-                    4: None,
-                    5: None,
-                    6: f"{self.user_role_name},{self.user_role_another_account_name}",
-                },
-                "projects": {0: None, 1: None, 2: None, 3: None, 4: None, 5: self.project.name, 6: None},
-                "teams": {0: self.team1.name, 1: None, 2: self.team2.name, 3: None, 4: None, 5: None, 6: None},
-                "phone_number": {0: None, 1: None, 2: None, 3: None, 4: None, 5: None, 6: None},
-                "editable_org_unit_types": {
-                    0: None,
-                    1: None,
-                    2: None,
-                    3: None,
-                    4: None,
-                    5: self.sub_unit_type.pk,
-                    6: None,
-                },
-            },
-        )
-
-    def test_profile_list_export_as_csv_multiple_teams(self):
-        self.maxDiff = None
-        multi_user = self.create_user_with_profile(username="multiteam", account=self.account)
-
-        multi_user.teams.set([self.team1, self.team2])
-
-        self.client.force_authenticate(self.jane)
-        response = self.client.get(reverse("profiles-export-csv", kwargs={"version": "v2"}))
-        self.assertEqual(response.status_code, 200)
-
-        csv_rows = self.assertCsvFileResponse(response, expected_name="users.csv", streaming=True, return_as_lists=True)
-
-        header = csv_rows[0]
-
-        username_idx = header.index("username")
-        teams_idx = header.index("teams")
-        user_row = next(row for row in csv_rows if row[username_idx] == "multiteam")
-
-        teams = sorted([self.team1, self.team2], key=lambda x: x.id)
-        expected_teams_value = f"{teams[0].name},{teams[1].name}"
-
-        self.assertEqual(user_row[teams_idx], expected_teams_value)
-
-    def test_profile_list_user_admin_ok(self):
-        """GET /profiles/ with auth (user has user admin permissions)"""
-        self.client.force_authenticate(self.jim)
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}))
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 7)
-
-    def test_profile_list_superuser_ok(self):
-        """GET /profiles/ with auth (superuser)"""
-        self.client.force_authenticate(self.john)
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}))
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 7)
-
-    def test_profile_list_user_manager_ok(self):
-        """GET /profiles/ with auth (superuser)"""
-        self.client.force_authenticate(self.jam)
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}))
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertEqual(response_data["count"], 7)
-        self.assertEqual(len(response_data["results"]), 7)
-        self.assertValidProfileListData(response_data, 7)
-
-    def test_profile_list_managed_user_only_superuser(self):
-        """GET /profiles/ with auth (superuser)"""
-        self.client.force_authenticate(self.john)
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}), {"managedUsersOnly": True})
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 7)
-
-    def test_profile_list_managed_user_only_user_admin(self):
-        """GET /profiles/ with auth (superuser)"""
-        self.client.force_authenticate(self.john)
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}), {"managedUsersOnly": True})
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 7)
-
-    def test_profile_list_managed_user_only_user_manager_no_org_unit(self):
-        """GET /profiles/ with auth (superuser)"""
-        self.client.force_authenticate(self.jam)
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}), {"managedUsersOnly": True})
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 6)
-
-    def test_profile_list_managed_user_only_user_manager_with_org_unit(self):
-        """GET /profiles/ with auth (superuser)"""
-        self.jam.iaso_profile.org_units.set([self.org_unit_from_parent_type.id])
-        self.jum.iaso_profile.org_units.set([self.child_org_unit.id])
-        self.client.force_authenticate(self.jam)
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}), {"managedUsersOnly": True})
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 2)
-
-    def test_profile_list_managed_user_only_user_regular_user(self):
-        """GET /profiles/ with auth (superuser)"""
-        self.client.force_authenticate(self.jane)
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}), {"managedUsersOnly": True})
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 0)
-
-    def test_create_profile_no_perm(self):
-        self.client.force_authenticate(self.jane)
-        data = {
-            "userName": "unittest_user_name",
-            "password": "unittest_password",
-            "firstName": "unittest_first_name",
-            "lastName": "unittest_last_name",
-        }
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-
-        self.assertJSONResponse(response, 403)
-
-    def test_create_user_with_user_roles_and_permissions(self):
-        self.client.force_authenticate(self.jim)
-        data = {
-            "userName": "unittest_user_name",
-            "password": "unittest_password",
-            "firstName": "unittest_first_name",
-            "lastName": "unittest_last_name",
-            "email": "john.doe@test.com",
-            "userPermissions": [CORE_FORMS_PERMISSION.codename],
-            "userRoles": [self.user_role.id],
-        }
-
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        response_data = self.assertJSONResponse(response, 201)
-
-        user_user_role = m.UserRole.objects.get(pk=response_data["userRoles"][0])
-        self.assertEqual(user_user_role.id, self.user_role.id)
-        self.assertEqual(user_user_role.group.name, self.group.name)
-
-        user = m.User.objects.get(username="unittest_user_name")
-        self.assertEqual(user.user_permissions.count(), 1)
-        self.assertEqual(user.user_permissions.first().codename, "iaso_forms")
-
-    def test_create_user_with_not_allowed_user_roles(self):
-        self.client.force_authenticate(self.jim)
-        data = {
-            "userName": "unittest_user_name",
-            "password": "unittest_password",
-            "firstName": "unittest_first_name",
-            "lastName": "unittest_last_name",
-            "email": "john.doe@test.com",
-            "userPermissions": [CORE_FORMS_PERMISSION.codename],
-            "userRoles": [self.user_role.id, self.user_role_another_account.id],
-        }
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        response_data = self.assertJSONResponse(response, 400)
-        self.assertHasError(response_data, "userRoles", "One or more user roles do not belong to the provided account.")
-
-    def test_create_profile_duplicate_user(self):
-        self.client.force_authenticate(self.jim)
-        data = {
-            "userName": "janedoe",
-            "password": "unittest_password",
-            "firstName": "unittest_first_name",
-            "lastName": "unittest_last_name",
-        }
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        response_data = self.assertJSONResponse(response, 400)
-        self.assertEqual(response_data["userName"], ["Username already exists for this account."])
-
-    def test_create_profile_duplicate_user_with_capital_letters(self):
-        self.client.force_authenticate(self.jim)
-        data = {
-            "userName": "JaNeDoE",
-            "password": "unittest_password",
-            "firstName": "unittest_first_name",
-            "lastName": "janedoe@test.com",
-        }
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        response_data = self.assertJSONResponse(response, 400)
-        self.assertEqual(response_data["userName"], ["Username already exists for this account."])
-
-    def test_create_profile_then_delete(self):
-        self.client.force_authenticate(self.jim)
-        data = {
-            "userName": "unittest_user_name",
-            "password": "unittest_password",
-            "firstName": "unittest_first_name",
-            "lastName": "unittest_last_name",
-            "email": "jane.doe@test.com",
-        }
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        response_data = self.assertJSONResponse(response, 201)
-
-        profile = m.Profile.objects.get(pk=response_data["id"])
-        user = profile.user
-        self.assertEqual(user.username, data["userName"])
-        self.assertEqual(user.first_name, data["firstName"])
-        self.assertQuerySetEqual(user.user_permissions.all(), [])
-        self.assertEqual(m.User.objects.filter(username=data["userName"]).count(), 1)
-        # check that we have copied the account from the creator account
-        self.assertEqual(profile.account, self.account)
-
-        profile_id = profile.id
-        user_id = user.id
-        response = self.client.delete(reverse("profiles-detail", kwargs={"pk": profile_id, "version": "v2"}))
-
-        self.assertJSONResponse(response, 204)
-        self.assertQuerySetEqual(m.User.objects.filter(id=user_id), [])
-        self.assertQuerySetEqual(m.Profile.objects.filter(id=profile_id), [])
-
-    def test_create_profile_with_org_units_and_perms(self):
-        self.client.force_authenticate(self.jim)
-        data = {
-            "userName": "unittest_user_name",
-            "password": "unittest_password",
-            "firstName": "unittest_first_name",
-            "lastName": "unittest_last_name",
-            "email": "john.doe@test.com",
-            "orgUnits": [self.org_unit_from_parent_type.id],
-            "userPermissions": [CORE_FORMS_PERMISSION.codename],
-            "editableOrgUnitTypeIds": [self.sub_unit_type.id],
-        }
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        response_data = self.assertJSONResponse(response, 201)
-
-        profile = m.Profile.objects.get(pk=response_data["id"])
-        self.assertEqual(profile.editable_org_unit_types.count(), 1)
-        self.assertEqual(profile.editable_org_unit_types.first(), self.sub_unit_type)
-
-        user = profile.user
-        self.assertEqual(user.username, data["userName"])
-        self.assertEqual(user.first_name, data["firstName"])
-
-        self.assertEqual(m.User.objects.filter(username=data["userName"]).count(), 1)
-        self.assertEqual(profile.account, self.account)
-
-        self.assertQuerySetEqual(
-            user.user_permissions.all(),
-            ["<Permission: iaso | core permission support | Formulaires>"],
-            transform=repr,
-        )
-        org_units = profile.org_units.all()
-        self.assertEqual(org_units.count(), 1)
-        self.assertEqual(org_units[0].name, "Corruscant Jedi Council")
-
-    def test_create_profile_with_color(self):
-        self.client.force_authenticate(self.jim)
-        color = "#123abc"
-        data = {
-            "userName": "color_user",
-            "password": "unittest_password",
-            "firstName": "color_first_name",
-            "lastName": "color_last_name",
-            "email": "color@example.com",
-            "color": color,
-        }
-
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        response_data = self.assertJSONResponse(response, 201)
-
-        self.assertEqual(response_data["color"], color.upper())
-
-        profile = m.Profile.objects.get(pk=response_data["id"])
-        self.assertEqual(profile.color, color.upper())
-
-    @override_settings(DEFAULT_FROM_EMAIL="sender@test.com", DNS_DOMAIN="iaso-test.bluesquare.org")
-    def test_create_profile_with_send_email(self):
-        self.client.force_authenticate(self.jim)
-        data = {
-            "userName": "userTest",
-            "password": "",
-            "firstName": "unittest_first_name",
-            "lastName": "unittest_last_name",
-            "sendEmailInvitation": True,
-            "email": "test@test.com",
-        }
-
-        with self.captureOnCommitCallbacks(execute=True) as callbacks:
-            response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-
-        response_data = self.assertJSONResponse(response, 201)
-        self.assertEqual(len(callbacks), 1)
-        self.assertEqual(len(mail.outbox), 1)
-        email = mail.outbox[0]
-
-        profile = Profile.objects.get(pk=response_data["id"])
-
-        # check plain text
-
-        # should not be translated
-        self.assertIn("Hello", email.body)
-        self.assertNotIn("Bonjour", email.body)
-
-        # checking variables in plain text
-        self.assertIn("iaso-test.bluesquare.org", email.body)
-        self.assertIn("http://", email.body)
-        self.assertIn("userTest", email.body)
-        self.assertIn(profile.account.name, email.body)
-        self.assertIn("http://iaso-test.bluesquare.org/reset-password-confirmation", email.body)
-
-        # check html
-
-        self.assertEqual(len(email.alternatives), 1)
-
-        html_content = email.alternatives[0][0]
-        self.assertIn("<p> Hello", html_content)
-
-        self.assertIn("iaso-test.bluesquare.org", html_content)
-        self.assertIn("http://", html_content)
-        self.assertIn("userTest", html_content)
-        self.assertIn(profile.account.name, html_content)
-        self.assertIn("http://iaso-test.bluesquare.org/reset-password-confirmation", html_content)
-
-        # check senders and receivers
-        self.assertEqual(email.from_email, "sender@test.com")
-        self.assertEqual(email.to, ["test@test.com"])
-
-        self.assertEqual(email.subject, "Set up a password for your new account on iaso-test.bluesquare.org")
-        self.assertEqual(email.from_email, "sender@test.com")
-        self.assertEqual(email.to, ["test@test.com"])
-        self.assertIn("http://iaso-test.bluesquare.org", email.body)
-        self.assertIn("The iaso-test.bluesquare.org Team.", email.body)
-
-    @override_settings(DEFAULT_FROM_EMAIL="sender@test.com", DNS_DOMAIN="iaso-test.bluesquare.org")
-    def test_create_profile_with_send_email_in_french(self):
-        self.client.force_authenticate(self.jim)
-        data = {
-            "userName": "userTest",
-            "password": "",
-            "firstName": "unittest_first_name",
-            "lastName": "unittest_last_name",
-            "sendEmailInvitation": True,
-            "email": "test@test.com",
-            "language": "fr",
-        }
-
-        with self.captureOnCommitCallbacks(execute=True) as callbacks:
-            response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-
-        result = self.assertJSONResponse(response, 201)
-        self.assertEqual(len(callbacks), 1)
-        self.assertEqual(len(mail.outbox), 1)
-        email = mail.outbox[0]
-        self.assertEqual(
-            email.subject, "Configurer un mot de passe pour votre nouveau compte sur iaso-test.bluesquare.org"
-        )
-
-        profile = Profile.objects.get(pk=result["id"])
-
-        # check plain text
-
-        # should be translated
-        self.assertNotIn("Hello", email.body)
-        self.assertIn("Bonjour", email.body)
-
-        # checking variables in plain text
-        self.assertIn("iaso-test.bluesquare.org", email.body)
-        self.assertIn("http://", email.body)
-        self.assertIn("userTest", email.body)
-        self.assertIn(profile.account.name, email.body)
-        self.assertIn("http://iaso-test.bluesquare.org/reset-password-confirmation", email.body)
-
-        # check html
-
-        self.assertEqual(len(email.alternatives), 1)
-
-        html_content = email.alternatives[0][0]
-        self.assertIn("<p>Bonjour", html_content)
-
-        self.assertIn("iaso-test.bluesquare.org", html_content)
-        self.assertIn("http://", html_content)
-        self.assertIn("userTest", html_content)
-        self.assertIn(profile.account.name, html_content)
-        self.assertIn("http://iaso-test.bluesquare.org/reset-password-confirmation", html_content)
-
-        # check senders and receivers
-        self.assertEqual(email.from_email, "sender@test.com")
-        self.assertEqual(email.to, ["test@test.com"])
-
-    def test_create_profile_with_no_password_and_not_send_email(self):
-        self.client.force_authenticate(self.jim)
-        data = {
-            "userName": "userTest",
-            "password": "",
-            "firstName": "unittest_first_name",
-            "lastName": "unittest_last_name",
-            "sendEmailInvitation": False,
-            "email": "test@test.com",
-        }
-
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        result = self.assertJSONResponse(response, 400)
-
-        self.assertHasError(result, "password", "This field is required.")
-
-    def test_create_profile_with_managed_geo_limit(self):
-        self.client.force_authenticate(self.user_managed_geo_limit)
-        data = {
-            "user_name": "unittest_user_name",
-            "password": "unittest_password",
-            "first_name": "unittest_first_name",
-            "last_name": "unittest_last_name",
-            "email": "john.doe@test.com",
-            "org_units": [self.child_org_unit.id],
-            "user_permissions": [CORE_FORMS_PERMISSION.codename],
-        }
-
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        response_data = self.assertJSONResponse(response, 201)
-
-        profile = m.Profile.objects.get(pk=response_data["id"])
-        user = profile.user
-        self.assertEqual(user.username, data["user_name"])
-        self.assertEqual(user.first_name, data["first_name"])
-
-        self.assertEqual(m.User.objects.filter(username=data["user_name"]).count(), 1)
-        self.assertEqual(profile.account, self.account)
-
-        self.assertQuerySetEqual(
-            user.user_permissions.all(),
-            ["<Permission: iaso | core permission support | Formulaires>"],
-            transform=repr,
-        )
-        org_units = profile.org_units.all()
-        self.assertEqual(org_units.count(), 1)
-        self.assertEqual(org_units[0].name, "Corruscant Jedi Council")
-
-    def test_create_profile_without_org_unit_with_managed_geo_limit(self):
-        self.client.force_authenticate(self.user_managed_geo_limit)
-        data = {
-            "user_name": "unittest_user_name",
-            "password": "unittest_password",
-            "first_name": "unittest_first_name",
-            "last_name": "unittest_last_name",
-            "email": "john.doe@test.com",
-            "user_permissions": [CORE_FORMS_PERMISSION.codename],
-        }
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-
-        self.assertJSONResponse(response, 403)
-
-    def test_create_user_is_atomic(self):
-        project_1 = m.Project.objects.create(name="Project 1", app_id="project.1", account=self.account)
-        project_2 = m.Project.objects.create(name="Project 2", app_id="project.2", account=self.account)
-
-        username = "john_doe"
-
-        user = self.user_managed_geo_limit
-        user.iaso_profile.projects.set([project_1])
-
-        self.client.force_authenticate(user)
-
-        self.assertEqual(get_user_model().objects.filter(username=username).count(), 0)
-
-        data = {
-            "user_name": username,
-            "password": "password",
-            "first_name": "John",
-            "last_name": "Doe",
-            "email": "john@doe.com",
-            "projects": [project_2.id],
-            "org_units": [self.org_unit_from_parent_type.id],
-        }
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-
-        self.assertJSONResponse(response, 403)
-
-        self.assertEqual(
-            response.data["detail"],
-            "Some projects are outside your scope.",
-        )
-
-        # If the creation is not successfully completed, no changes should be committed to the database.
-        self.assertEqual(get_user_model().objects.filter(username=username).count(), 0)
-
-    def test_delete_profile_no_perm(self):
-        self.client.force_authenticate(self.jane)
-        response = self.client.delete(reverse("profiles-detail", kwargs={"pk": 1, "version": "v2"}))
-
-        self.assertJSONResponse(response, 403)
-
-    def test_profile_error_dhis2_constraint(self):
-        # Test for regression of IA-1249
-        self.client.force_authenticate(self.jim)
-        data = {"user_name": "unittest_user1", "password": "unittest_password", "dhis2_id": ""}
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        self.assertJSONResponse(response, 201)
-
-        data = {"user_name": "unittest_user2", "password": "unittest_password", "dhis2_id": ""}
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-
-        self.assertJSONResponse(response, 201)
-        profile1 = m.Profile.objects.get(user__username="unittest_user1")
-        profile2 = m.Profile.objects.get(user__username="unittest_user2")
-        self.assertNotEqual(profile1.account_id, None)
-        self.assertEqual(profile2.account_id, profile1.account_id)
-        self.assertEqual(profile2.dhis2_id, None)
-
-        data = {"user_name": "unittest_user2", "password": "unittest_password", "dhis2_id": "", "first_name": "test"}
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile2.id, "version": "v2"}), data=data, format="json"
-        )
-        self.assertJSONResponse(response, 200)
-        profile2.refresh_from_db()
-        self.assertEqual(profile2.dhis2_id, None)
-
-        data = {"user_name": "unittest_user2", "password": "unittest_password", "dhis2_id": "test_dhis2_id"}
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile2.id, "version": "v2"}), data=data, format="json"
-        )
-        self.assertJSONResponse(response, 200)
-        profile2.refresh_from_db()
-        self.assertEqual(profile2.dhis2_id, "test_dhis2_id")
-
-    def test_account_feature_flags_is_included(self):
-        aff = m.AccountFeatureFlag.objects.create(code="shape", name="Can edit shape")
-        m.AccountFeatureFlag.objects.create(code="not-used", name="this is not used")
-        self.client.force_authenticate(self.jane)
-
-        # no feature flag at first
-        response = self.client.get(reverse("profiles-detail", kwargs={"pk": "me", "version": "v2"}))
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertIn("account", response_data)
-        self.assertEqual(response_data["account"]["featureFlags"], [])
-
-        # add a feature flags
-        self.account.feature_flags.add(aff)
-
-        response = self.client.get(reverse("profiles-detail", kwargs={"pk": "me", "version": "v2"}))
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertIn("account", response_data)
-
-        self.assertEqual(response_data["account"]["featureFlags"], ["shape"])
-
-        # remove feature flags
-        self.account.feature_flags.remove(aff)
-        response = self.client.get(reverse("profiles-detail", kwargs={"pk": "me", "version": "v2"}))
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertIn("account", response_data)
-
-        self.assertEqual(response_data["account"]["featureFlags"], [])
-
-    def test_search_user_by_has_email(self):
-        self.client.force_authenticate(self.jane)
-
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}), {"hasEmail": True})
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertEqual(len(response_data["results"]), 0)
-        self.assertTrue(all(x["email"] for x in response_data["results"]))
-
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}), {"hasEmail": False})
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertEqual(len(response_data["results"]), 7)
-        self.assertFalse(all(x["email"] for x in response_data["results"]))
-
-    def test_search_user_by_permissions(self):
-        self.client.force_authenticate(self.jane)
-
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}), {"permissions": "iaso_users"})
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertEqual(len(response_data["results"]), 1)
-        self.assertEqual(response_data["results"][0]["userName"], "jim")
-
-    def test_search_user_by_org_units(self):
-        self.client.force_authenticate(self.jane)
-        self.jane.iaso_profile.org_units.set([self.org_unit_from_parent_type])
-
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v2"}),
-            {"location": self.org_unit_from_parent_type.pk, "limit": 100},
-        )
-
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 2)
-        self.assertEqual(response_data["results"][0]["userName"], "janedoe")
-
-    def test_search_user_by_org_units_type(self):
-        self.client.force_authenticate(self.jane)
-        self.jane.iaso_profile.org_units.set([self.org_unit_from_parent_type])
-
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v2"}),
-            {"orgUnitTypes": self.parent_org_unit_type.pk, "limit": 100},
-        )
-
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 2)
-        self.assertEqual(response_data["results"][0]["userName"], "janedoe")
-
-    def test_search_user_by_children_ou(self):
-        self.client.force_authenticate(self.jane)
-        self.jane.iaso_profile.org_units.set([self.child_org_unit])
-
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v2"}),
-            {"location": self.org_unit_from_parent_type.pk, "ouParent": False, "ouChildren": True},
-        )
-
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 2)
-
-        self.assertEqual(response_data["results"][0]["userName"], "janedoe")
-
-    def test_search_user_by_parent_ou(self):
-        self.client.force_authenticate(self.jane)
-        self.jane.iaso_profile.org_units.set([self.org_unit_from_parent_type])
-
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v2"}),
-            {"location": self.child_org_unit.pk, "ouParent": True, "ouChildren": False, "limit": 100},
-        )
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 2)
-        self.assertEqual(response_data["results"][0]["userName"], "janedoe")
-
-    def test_list_by_ids(self):
-        self.client.force_authenticate(self.jane)
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v2"}), {"ids": f"{self.jane.id},{self.jim.id}"}
-        )
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 2)
-        self.assertEqual(response_data["results"][0]["userName"], "janedoe")
-        self.assertEqual(response_data["results"][1]["userName"], "jim")
-
-    def test_search_by_profile_ids(self):
-        self.client.force_authenticate(self.jane)
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v2"}),
-            {"search": f"ids:{self.jane.iaso_profile.id},{self.jim.iaso_profile.id}", "order": "id"},
-        )
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 2)
-        self.assertEqual(response_data["results"][0]["userName"], "janedoe")
-        self.assertEqual(response_data["results"][1]["userName"], "jim")
-
-    def test_search_by_dhis2_id(self):
-        self.client.force_authenticate(self.jane)
-        mydhis2_id = "mydhis2id"
-
-        self.jim.iaso_profile.dhis2_id = mydhis2_id
-        self.jim.iaso_profile.save()
-
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}), {"search": f"refs:{mydhis2_id}"})
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 1)
-        self.assertEqual(response_data["results"][0]["userName"], "jim")
-
-    def test_search_by_teams(self):
-        self.client.force_authenticate(self.jane)
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v2"}), {"teams": f"{self.team1.pk},{self.team2.pk}"}
-        )
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 2)
-        user_names = [item["userName"] for item in response_data["results"]]
-        self.assertIn("janedoe", user_names)
-        self.assertIn("jim", user_names)
-
-    def test_user_with_managed_permission_can_update_profile_of_user_in_sub_org_unit(self):
-        self.jam.iaso_profile.org_units.set([self.org_unit_from_parent_type.id])
-        self.jum.iaso_profile.org_units.set([self.child_org_unit.id])
-        self.client.force_authenticate(self.jam)
-        jum = Profile.objects.get(user=self.jum)
-        data = {
-            "userName": "unittest_user_name",
-            "password": "unittest_password",
-            "firstName": "unittest_first_name",
-            "lastName": "unittest_last_name",
-            "userPermissions": [CORE_FORMS_PERMISSION.codename, CORE_USERS_MANAGED_PERMISSION.codename],
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": jum.pk, "version": "v2"}), data=data, format="json"
-        )
-        self.assertJSONResponse(response, 200)
-
-    def test_user_with_managed_permission_without_location_can_update_profile_of_user_in_whole_pyramid(self):
-        self.jum.iaso_profile.org_units.set([self.child_org_unit.id])
-        self.client.force_authenticate(self.jam)
-        jum = Profile.objects.get(user=self.jum)
-        data = {
-            "user_name": "unittest_user_name",
-            "password": "unittest_password",
-            "first_name": "unittest_first_name",
-            "last_name": "unittest_last_name",
-            "org_units": [self.org_unit_from_parent_type.id],
-            "user_permissions": [CORE_FORMS_PERMISSION.codename, CORE_USERS_MANAGED_PERMISSION.codename],
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": jum.id, "version": "v2"}), data=data, format="json"
-        )
-        jum.refresh_from_db()
-        self.assertJSONResponse(response, 200)
-        self.assertEqual(list(jum.org_units.values_list("id", flat=True)), [self.org_unit_from_parent_type.id])
-
-    def test_user_with_managed_permission_cannot_grant_user_admin_permission(self):
-        self.jam.iaso_profile.org_units.set([self.org_unit_from_parent_type.id])
-        self.jum.iaso_profile.org_units.set([self.child_org_unit.id])
-        self.client.force_authenticate(self.jam)
-        jum = Profile.objects.get(user=self.jum)
-        data = {
-            "user_name": "jum",
-            "user_permissions": [
-                CORE_FORMS_PERMISSION.codename,
-                CORE_USERS_MANAGED_PERMISSION.codename,
-                CORE_USERS_ADMIN_PERMISSION.codename,
-            ],
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": jum.id, "version": "v2"}), data=data, format="json"
-        )
-        self.assertJSONResponse(response, 403)
-
-    def test_user_with_managed_permission_cannot_grant_user_admin_permission_through_user_roles(self):
-        group = Group.objects.create(name="admin")
-        group.permissions.set([Permission.objects.get(codename=CORE_USERS_ADMIN_PERMISSION.codename)])
-        role = m.UserRole.objects.create(account=self.account, group=group)
-        self.jam.iaso_profile.org_units.set([self.org_unit_from_parent_type.id])
-        self.jum.iaso_profile.org_units.set([self.child_org_unit.id])
-        self.client.force_authenticate(self.jam)
-        jum = Profile.objects.get(user=self.jum)
-        data = {
-            "user_name": "jum",
-            "user_roles": [role.id],
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": jum.id, "version": "v2"}), data=data, format="json"
-        )
-        self.assertJSONResponse(response, 403)
-
-    def test_user_with_managed_permission_can_grant_user_roles(self):
-        group = Group.objects.create(name="admin")
-        group.permissions.set([Permission.objects.get(codename=CORE_FORMS_PERMISSION.codename)])
-        role = m.UserRole.objects.create(account=self.account, group=group)
-        self.jam.iaso_profile.org_units.set([self.org_unit_from_parent_type.id])
-        self.jum.iaso_profile.org_units.set([self.child_org_unit.id])
-        self.client.force_authenticate(self.jam)
-        jum = Profile.objects.get(user=self.jum)
-        data = {
-            "user_name": "jum",
-            "user_roles": [role.id],
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": jum.id, "version": "v2"}), data=data, format="json"
-        )
-        self.assertJSONResponse(response, 200)
-
-    def test_user_with_managed_permission_can_assign_org_unit_within_their_health_pyramid(self):
-        self.jam.iaso_profile.org_units.set([self.org_unit_from_parent_type.id])
-        self.jum.iaso_profile.org_units.set([self.child_org_unit.id])
-        self.client.force_authenticate(self.jam)
-        jum = Profile.objects.get(user=self.jum)
-        data = {
-            "user_name": "jum",
-            "org_units": [self.org_unit_from_parent_type.id],
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": jum.id, "version": "v2"}), data=data, format="json"
-        )
-        self.assertJSONResponse(response, 200)
-
-    def test_user_with_managed_permission_can_assign_org_unit_within_their_health_pyramid_with_existing_ones_outside(
-        self,
-    ):
-        self.jam.iaso_profile.org_units.set([self.org_unit_from_parent_type.id])
-        self.jum.iaso_profile.org_units.set([self.child_org_unit.id, self.org_unit_from_sub_type])
-        self.client.force_authenticate(self.jam)
-        jum = Profile.objects.get(user=self.jum)
-        data = {
-            "user_name": "jum",
-            "org_units": [self.org_unit_from_parent_type.id, self.org_unit_from_sub_type.id],
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": jum.id, "version": "v2"}), data=data, format="json"
-        )
-        self.assertJSONResponse(response, 200)
-
-    def test_user_with_managed_permission_cannot_assign_org_unit_outside_of_their_health_pyramid(self):
-        self.jam.iaso_profile.org_units.set([self.org_unit_from_parent_type.id])
-        self.jum.iaso_profile.org_units.set([self.child_org_unit.id])
-        self.client.force_authenticate(self.jam)
-        jum = Profile.objects.get(user=self.jum)
-        data = {
-            "user_name": "jum",
-            "org_units": [self.org_unit_from_sub_type.id],
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": jum.id, "version": "v2"}), data=data, format="json"
-        )
-        self.assertJSONResponse(response, 403)
-
-    def test_user_with_managed_permission_cannot_update_profile_of_user_not_in_sub_org_unit(self):
-        self.jam.iaso_profile.org_units.set([self.org_unit_from_parent_type.id])
-        self.client.force_authenticate(self.jam)
-        data = {
-            "user_name": "unittest_user_name",
-            "password": "unittest_password",
-            "first_name": "unittest_first_name",
-            "last_name": "unittest_last_name",
-        }
-        jum = Profile.objects.get(user=self.jum)
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": jum.id, "version": "v2"}), data=data, format="json"
-        )
-        self.assertJSONResponse(response, 403)
-
-    def test_user_with_managed_permission_can_update_profile_if_not_themselves_in_sub_org_unit(self):
-        self.jum.iaso_profile.org_units.set([self.org_unit_from_parent_type.id])
-        self.client.force_authenticate(self.jam)
-        data = {
-            "user_name": "unittest_user_name",
-            "password": "unittest_password",
-            "first_name": "unittest_first_name",
-            "last_name": "unittest_last_name",
-        }
-        jum = Profile.objects.get(user=self.jum)
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": jum.id, "version": "v2"}), data=data, format="json"
-        )
-        self.assertJSONResponse(response, 200)
-
-    def test_update_user_should_fail_when_assigning_an_org_unit_outside_of_the_author_own_health_pyramid(self):
-        user = self.jam
-        user.iaso_profile.org_units.set([self.org_unit_from_sub_type])
-
-        self.assertTrue(user.has_perm(CORE_USERS_MANAGED_PERMISSION.full_name()))
-        self.assertFalse(user.has_perm(CORE_USERS_ADMIN_PERMISSION.full_name()))
-
-        profile_to_modify = Profile.objects.get(user=self.jum)
-        profile_to_modify.org_units.set([self.org_unit_from_sub_type])
-
-        self.client.force_authenticate(user)
-        data = {
-            "user_name": "new_user_name",
-            "org_units": [self.org_unit_from_parent_type.pk, self.another_org_unit.pk],
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile_to_modify.id, "version": "v2"}), data=data, format="json"
-        )
-        response_data = self.assertJSONResponse(response, 403)
-        self.assertEqual(
-            response_data["detail"],
-            (
-                f"User with {CORE_USERS_MANAGED_PERMISSION} cannot assign an OrgUnit outside "
-                f"of their own health pyramid. Trying to assign {self.another_org_unit.pk}."
-            ),
-        )
-
-    def test_update_user_should_succeed_with_restricted_editable_org_unit_types_when_modifying_another_field(self):
-        """
-        A user restricted to a given OrgUnitType should be able to edit another user as long as
-        he's not modifying the `org_units` or `editable_org_unit_type_ids` fields.
-        """
-        user = self.jam
-
-        self.assertTrue(user.has_perm(CORE_USERS_MANAGED_PERMISSION.full_name()))
-        self.assertFalse(user.has_perm(CORE_USERS_ADMIN_PERMISSION.full_name()))
-
-        user.iaso_profile.org_units.set([self.org_unit_from_parent_type])
-        user.iaso_profile.editable_org_unit_types.set(
-            # Only org units of this type is now writable.
-            [self.sub_unit_type]
-        )
-
-        jum_profile = Profile.objects.get(user=self.jum)
-        jum_profile.org_units.set([self.org_unit_from_parent_type])
-        jum_profile.editable_org_unit_types.set(
-            # Only org units of this type is now writable.
-            [self.parent_org_unit_type]
-        )
-
-        self.client.force_authenticate(user)
-
-        data = {
-            "userName": "new_user_name",
-            "orgUnits": [self.org_unit_from_parent_type.id],
-            "editableOrgUnitTypeIds": [self.parent_org_unit_type.id],
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": jum_profile.id, "version": "v2"}), data=data, format="json"
-        )
-        self.assertJSONResponse(response, 200)
-        self.jum.refresh_from_db()
-        self.assertEqual(self.jum.username, "new_user_name")
-
-    def test_user_with_managed_permission_cannot_create_users(self):
-        self.jam.iaso_profile.org_units.set([self.org_unit_from_parent_type.id])
-        self.client.force_authenticate(self.jam)
-        data = {
-            "userName": "unittest_user_name",
-            "password": "unittest_password",
-            "firstName": "unittest_first_name",
-            "lastName": "unittest_last_name",
-        }
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        self.assertJSONResponse(response, 403)
-
-    def test_user_with_managed_permission_cannot_delete_users(self):
-        self.jam.iaso_profile.org_units.set([self.org_unit_from_parent_type.id])
-        self.client.force_authenticate(self.jam)
-        jum = Profile.objects.get(user=self.jum)
-        response = self.client.delete(reverse("profiles-detail", kwargs={"pk": jum.id, "version": "v2"}))
-        self.assertJSONResponse(response, 403)
-
-    def test_user_with_managed_permission_cannot_update_from_unmanaged_org_unit(self):
-        self.jam.iaso_profile.org_units.set([self.child_org_unit.id])
-        self.jum.iaso_profile.org_units.set([self.org_unit_from_sub_type.id])
-        self.client.force_authenticate(self.jam)
-        data = {
-            "userName": "unittest_user_name",
-            "password": "unittest_password",
-            "firstName": "unittest_first_name",
-            "lastName": "unittest_last_name",
-        }
-        jum = Profile.objects.get(user=self.jum)
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": jum.id, "version": "v2"}), data=data, format="json"
-        )
-        self.assertJSONResponse(response, 403)
-
-    def test_update_user_add_phone_number(self):
-        self.jam.iaso_profile.org_units.set([self.org_unit_from_parent_type.id])
-        self.client.force_authenticate(self.john)
-        jum = Profile.objects.get(user=self.jum)
-        data = {
-            "userName": "unittest_user_name",
-            "password": "unittest_password",
-            "firstName": "unittest_first_name",
-            "lastName": "unittest_last_name",
-            "phoneNumber": "+32477123456",
-            "countryCode": "be",
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": jum.id, "version": "v2"}), data=data, format="json"
-        )
-        self.assertJSONResponse(response, 200)
-        updated_jum = Profile.objects.get(user=self.jum)
-        self.assertEqual(updated_jum.phone_number.as_e164, "+32477123456")
-
-        res = self.client.get(reverse("profiles-detail", kwargs={"pk": jum.id, "version": "v2"}))
-        response_data = self.assertJSONResponse(res, 200)
-        self.assertEqual(response_data["phoneNumber"], "+32477123456")
-
-    def test_update_profile_color(self):
-        self.client.force_authenticate(self.john)
-        profile = Profile.objects.get(user=self.jim)
-        original_first_name = profile.user.first_name
-        original_last_name = profile.user.last_name
-        original_email = profile.user.email
-        new_color = "#A1B2C3"
-        data = {"color": new_color}
-
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile.id, "version": "v2"}), data=data, format="json"
-        )
-
-        response_data = self.assertJSONResponse(response, 200)
-        profile.refresh_from_db()
-        profile.user.refresh_from_db()
-        self.assertEqual(profile.color, new_color)
-        self.assertEqual(response_data["color"], new_color)
-        self.assertEqual(profile.user.first_name, original_first_name)
-        self.assertEqual(profile.user.last_name, original_last_name)
-        self.assertEqual(profile.user.email, original_email)
-
-    def test_update_user_with_malformed_phone_number(self):
-        user = self.jam
-        profile_to_edit = Profile.objects.get(user=self.jum)
-
-        self.client.force_authenticate(user)
-
-        data = {
-            "userName": "new_name",
-            "phoneNumber": "not_a_phone_number",
-            "countryCode": "",
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile_to_edit.id, "version": "v2"}), data=data, format="json"
-        )
-        response_data = self.assertJSONResponse(response, 400)
-        self.assertHasError(response_data, "phoneNumber", "Both phone number and country code must be provided")
-
-        data = {
-            "userName": "new_name",
-            "phoneNumber": "not_a_phone_number",
-            "countryCode": "US",
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile_to_edit.id, "version": "v2"}), data=data, format="json"
-        )
-        response_data = self.assertJSONResponse(response, 400)
-        self.assertHasError(response_data, "phoneNumber", "Enter a valid phone number.")
-
-        data = {
-            "userName": "new_name",
-            "phoneNumber": "03666666",
-            "countryCode": "FR",
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile_to_edit.id, "version": "v2"}), data=data, format="json"
-        )
-        response_data = self.assertJSONResponse(response, 400)
-        self.assertHasError(response_data, "phoneNumber", "Enter a valid phone number.")
-
-        data = {
-            "userName": "new_name",
-            "phoneNumber": "0612345678",
-            "countryCode": "FR",
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile_to_edit.id, "version": "v2"}), data=data, format="json"
-        )
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertEqual(response_data["phoneNumber"], "+33612345678")
-
-        data = {"userName": "new_name"}
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile_to_edit.id, "version": "v2"}), data=data, format="json"
-        )
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertEqual(response_data["phoneNumber"], "+33612345678")
-
-        data = {"userName": "new_name", "phoneNumber": ""}
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile_to_edit.id, "version": "v2"}), data=data, format="json"
-        )
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertEqual(response_data["phoneNumber"], None)
-
-    def test_update_user_with_projects_restrictions(self):
-        new_project_1 = m.Project.objects.create(name="New project 1", app_id="new.project.1", account=self.account)
-        new_project_2 = m.Project.objects.create(name="New project 2", app_id="new.project.2", account=self.account)
-        profile_to_edit = Profile.objects.get(user=self.jum)
-        profile_to_edit.projects.clear()
-        user = self.jam
-        self.client.force_authenticate(user)
-        self.assertEqual(user.iaso_profile.projects.count(), 0)
-        self.assertEqual(profile_to_edit.projects.count(), 0)
-
-        # A user without `projects` restrictions can set any project.
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile_to_edit.id, "version": "v2"}),
-            data={
-                "user_name": "jum_new_user_name",
-                "projects": [self.project.id],
-            },
-            format="json",
-        )
-        self.assertJSONResponse(response, 200)
-        profile_to_edit.refresh_from_db()
-        self.assertEqual(profile_to_edit.projects.count(), 1)
-        self.assertEqual(profile_to_edit.projects.first(), self.project)
-        self.assertEqual(profile_to_edit.user.username, "jum_new_user_name")
-
-        # A user with `projects` restrictions can edit a user with the same `projects` restrictions.
-        user.iaso_profile.projects.clear()
-        profile_to_edit.projects.clear()
-        del user.iaso_profile.projects_ids  # Refresh cached property.
-        user.iaso_profile.projects.set([self.project, new_project_1, new_project_2])
-        profile_to_edit.projects.set([self.project, new_project_1, new_project_2])
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile_to_edit.id, "version": "v2"}),
-            data={
-                "user_name": "jum_new_user_name",
-                "projects": [self.project.id, new_project_1.id],
-            },
-            format="json",
-        )
-        self.assertJSONResponse(response, 200)
-        profile_to_edit.refresh_from_db()
-        self.assertEqual(profile_to_edit.projects.count(), 2)
-
-        # A user with `projects` restrictions cannot edit a user who has broader access to projects.
-        user.iaso_profile.projects.clear()
-        profile_to_edit.projects.clear()
-        del user.iaso_profile.projects_ids  # Refresh cached property.
-
-        user.iaso_profile.projects.set([self.project])
-        profile_to_edit.projects.set([self.project, new_project_1, new_project_2])
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile_to_edit.id, "version": "v2"}),
-            data={
-                "user_name": "jum_new_user_name",
-                "projects": [self.project.id],
-            },
-            format="json",
-        )
-        response_data = self.assertJSONResponse(response, 403)
-        self.assertEqual(
-            response_data["detail"],
-            "You cannot edit a user who has broader access to projects.",
-        )
-
-        # A user with `projects` restrictions can edit a user who has narrower access to projects.
-        user.iaso_profile.projects.clear()
-        profile_to_edit.projects.clear()
-        del user.iaso_profile.projects_ids  # Refresh cached property.
-        user.iaso_profile.projects.set([self.project, new_project_1])
-        profile_to_edit.projects.set([self.project])
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile_to_edit.id, "version": "v2"}),
-            data={
-                "user_name": "jum_new_user_name",
-                "projects": [new_project_1.id],
-            },
-            format="json",
-        )
-        self.assertJSONResponse(response, 200)
-        profile_to_edit.refresh_from_db()
-        self.assertEqual(profile_to_edit.projects.count(), 1)
-        self.assertEqual(profile_to_edit.projects.first(), new_project_1)
-        self.assertEqual(profile_to_edit.user.username, "jum_new_user_name")
-
-        # A user with `projects` restrictions cannot create a user without restrictions.
-        user.iaso_profile.projects.clear()
-        profile_to_edit.projects.clear()
-        del user.iaso_profile.projects_ids  # Refresh cached property.
-        user.iaso_profile.projects.set([self.project])
-        self.assertEqual(user.iaso_profile.projects.count(), 1)
-        profile_to_edit.projects.clear()
-        self.assertEqual(profile_to_edit.projects.count(), 0)
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile_to_edit.id, "version": "v2"}),
-            data={"user_name": "jum_new_user_name", "projects": []},
-            format="json",
-        )
-        response_data = self.assertJSONResponse(response, 403)
-        self.assertEqual(
-            response_data["detail"],
-            "You must specify which projects are authorized for this user.",
-        )
-
-        # A user with `projects` restrictions cannot assign projects outside his range.
-        user.iaso_profile.projects.clear()
-        profile_to_edit.projects.clear()
-        del user.iaso_profile.projects_ids  # Refresh cached property.
-        user.iaso_profile.projects.set([self.project])
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile_to_edit.id, "version": "v2"}),
-            data={
-                "user_name": "jum_new_user_name",
-                "projects": [new_project_2.id],
-            },
-            format="json",
-        )
-        response_data = self.assertJSONResponse(response, 403)
-        self.assertEqual(
-            response_data["detail"],
-            "Some projects are outside your scope.",
-        )
-
-        # An "admin" user with `projects` restrictions can assign projects outside his range.
-        user.user_permissions.add(Permission.objects.get(codename=CORE_USERS_ADMIN_PERMISSION.codename))
-        del user._perm_cache
-        del user._user_perm_cache
-        self.assertTrue(user.has_perm(CORE_USERS_ADMIN_PERMISSION.full_name()))
-        user.iaso_profile.projects.set([self.project])
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile_to_edit.id, "version": "v2"}),
-            data={
-                "user_name": "jum_new_user_name",
-                "projects": [new_project_2.id],
-            },
-            format="json",
-        )
-        self.assertJSONResponse(response, 200)
-
-    def test_admin_should_be_able_to_bypass_projects_restrictions_for_himself(self):
-        """
-        An admin with `projects` restrictions should be able to assign himself to any project.
-        """
-        project_1 = m.Project.objects.create(name="Project 1", app_id="project.1", account=self.account)
-        project_2 = m.Project.objects.create(name="Project 2", app_id="project.2", account=self.account)
-
-        user_admin = self.jim
-        self.assertFalse(user_admin.has_perm(CORE_USERS_MANAGED_PERMISSION.full_name()))
-        self.assertTrue(user_admin.has_perm(CORE_USERS_ADMIN_PERMISSION.full_name()))
-
-        profile_to_edit = user_admin.iaso_profile
-        profile_to_edit.projects.set([project_1])
-        self.assertEqual(profile_to_edit.projects.count(), 1)
-        self.assertEqual(profile_to_edit.projects.first(), project_1)
-
-        self.client.force_authenticate(user_admin)
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": profile_to_edit.id, "version": "v2"}),
-            data={
-                "userName": user_admin.username,
-                "projects": [project_2.id],
-            },
-            format="json",
-        )
-        self.assertJSONResponse(response, 200)
-
-        profile_to_edit.refresh_from_db()
-        self.assertEqual(profile_to_edit.projects.count(), 1)
-        self.assertEqual(profile_to_edit.projects.first(), project_2)
-
-    def test_log_on_user_create(self):
-        self.client.force_authenticate(self.john)
-
-        data = self.get_new_user_data()
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        response_data = self.assertJSONResponse(response, 201)
-        new_profile_id = response_data["id"]
-
-        new_user_id = Profile.objects.get(pk=new_profile_id).user_id
-
-        response = self.client.get(
-            reverse("logs-list"),
-            {
-                "contentType": "iaso.profile",
-                "fields": ",".join(["past_value", "new_value"]),
-                "objectId": new_profile_id,
-            },
-        )
-        response_data = self.assertJSONResponse(response, 200)
-        logs = response_data["list"]
-        log = logs[0]
-
-        try:
-            jsonschema.validate(instance=log, schema=PROFILE_LOG_SCHEMA)
-        except jsonschema.exceptions.ValidationError as ex:
-            self.fail(msg=str(ex))
-
-        self.assertEqual(log["past_value"], [])
-        self.assertEqual(log["new_value"][0]["pk"], new_profile_id)
-        new_profile = log["new_value"][0]["fields"]
-        self.assertEqual(new_profile["user"], new_user_id)
-        self.assertEqual(new_profile["username"], data["userName"])
-        self.assertEqual(new_profile["first_name"], data["firstName"])
-        self.assertEqual(new_profile["last_name"], data["lastName"])
-        self.assertEqual(new_profile["email"], data["email"])
-        self.assertEqual(new_profile["organization"], data["organization"])
-        self.assertEqual(len(new_profile["user_permissions"]), 1)
-        self.assertEqual(new_profile["user_permissions"], data["userPermissions"])
-        self.assertTrue(new_profile["password_updated"])
-        self.assertNotIn("password", new_profile.keys())
-
-        self.assertEqual(new_profile["dhis2_id"], data["dhis2Id"])
-        self.assertEqual(new_profile["language"], data["language"])
-        self.assertEqual(new_profile["home_page"], data["homePage"])
-        self.assertEqual(new_profile["phone_number"], f"{data['phoneNumber']}")
-        self.assertEqual(len(new_profile["org_units"]), 1)
-        self.assertIn(self.org_unit_from_parent_type.id, new_profile["org_units"])
-        self.assertEqual(len(new_profile["user_roles"]), 1)
-        self.assertIn(self.user_role.id, new_profile["user_roles"])
-        self.assertEqual(len(new_profile["projects"]), 1)
-        self.assertIn(self.project.id, new_profile["projects"])
-
-    def test_log_on_user_delete(self):
-        self.client.force_authenticate(self.john)
-        data = self.get_new_user_data()
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        response_data = self.assertJSONResponse(response, 201)
-        new_profile_id = response_data["id"]
-        # todo : really needed in the response ?
-        # new_user_id = response_data["user_id"]
-        new_user_id = Profile.objects.get(pk=new_profile_id).user_id
-
-        response = self.client.delete(reverse("profiles-detail", kwargs={"pk": new_profile_id, "version": "v2"}))
-        self.assertJSONResponse(response, 204)
-
-        response = self.client.get(
-            reverse("logs-list"),
-            {
-                "contentType": "iaso.profile",
-                "fields": ",".join(["past_value", "new_value"]),
-                "objectId": new_profile_id,
-            },
-        )
-        response_data = self.assertJSONResponse(response, 200)
-        logs = response_data["list"]
-        log = logs[0]
-
-        try:
-            jsonschema.validate(instance=log, schema=PROFILE_LOG_SCHEMA)
-        except jsonschema.exceptions.ValidationError as ex:
-            self.fail(msg=str(ex))
-
-        self.assertEqual(log["new_value"][0]["pk"], new_profile_id)
-        new_profile = log["new_value"][0]["fields"]
-        self.assertEqual(new_profile["user"], new_user_id)
-        self.assertEqual(new_profile["username"], data["userName"])
-        self.assertEqual(new_profile["first_name"], data["firstName"])
-        self.assertEqual(new_profile["last_name"], data["lastName"])
-        self.assertEqual(new_profile["email"], data["email"])
-        self.assertEqual(len(new_profile["user_permissions"]), 1)
-        self.assertEqual(new_profile["user_permissions"], data["userPermissions"])
-        self.assertIsNotNone(new_profile["deleted_at"])
-        self.assertFalse(new_profile["password_updated"])
-        self.assertNotIn("password", new_profile.keys())
-
-        self.assertEqual(new_profile["dhis2_id"], data["dhis2Id"])
-        self.assertEqual(new_profile["language"], data["language"])
-        self.assertEqual(new_profile["home_page"], data["homePage"])
-        self.assertEqual(new_profile["phone_number"], data["phoneNumber"])
-        self.assertEqual(len(new_profile["org_units"]), 1)
-        self.assertIn(self.org_unit_from_parent_type.id, new_profile["org_units"])
-        self.assertEqual(len(new_profile["user_roles"]), 1)
-        self.assertIn(self.user_role.id, new_profile["user_roles"])
-        self.assertEqual(len(new_profile["projects"]), 1)
-        self.assertIn(self.project.id, new_profile["projects"])
-        self.assertIsNotNone(new_profile["deleted_at"])
-
-        self.assertEqual(log["past_value"][0]["pk"], new_profile_id)
-        past_profile = log["past_value"][0]["fields"]
-        self.assertEqual(past_profile["user"], new_user_id)
-        self.assertEqual(past_profile["username"], data["userName"])
-        self.assertEqual(past_profile["first_name"], data["firstName"])
-        self.assertEqual(past_profile["last_name"], data["lastName"])
-        self.assertEqual(past_profile["email"], data["email"])
-        self.assertEqual(len(past_profile["user_permissions"]), 1)
-        self.assertEqual(past_profile["user_permissions"], data["userPermissions"])
-        self.assertIsNone(past_profile["deleted_at"])
-        self.assertNotIn("password", past_profile.keys())
-
-        self.assertEqual(past_profile["dhis2_id"], data["dhis2Id"])
-        self.assertEqual(past_profile["language"], data["language"])
-        self.assertEqual(past_profile["home_page"], data["homePage"])
-        self.assertEqual(past_profile["phone_number"], data["phoneNumber"])
-        self.assertEqual(len(past_profile["org_units"]), 1)
-        self.assertIn(self.org_unit_from_parent_type.id, past_profile["org_units"])
-        self.assertEqual(len(past_profile["user_roles"]), 1)
-        self.assertIn(self.user_role.id, past_profile["user_roles"])
-        self.assertEqual(len(past_profile["projects"]), 1)
-        self.assertIn(self.project.id, past_profile["projects"])
-        self.assertIsNone(past_profile["deleted_at"])
-
-    def test_log_on_user_update(self):
-        self.client.force_authenticate(self.john)
-        data = self.get_new_user_data()
-        response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        response_data = self.assertJSONResponse(response, 201)
-        new_profile_id = response_data["id"]
-        new_user_id = Profile.objects.get(pk=new_profile_id).user_id
-        new_user_name = Profile.objects.get(pk=new_profile_id).user.username
-
-        new_data = {
-            "id": new_profile_id,
-            "orgUnits": [self.org_unit_from_parent_type.id, self.org_unit_from_sub_type.id],
-            "language": "en",
-            "password": "yolo",
-            "homePage": "/orgunits/list",
-            "organization": "Bluesquare",
-            "userPermissions": [CORE_ORG_UNITS_READ_PERMISSION.codename, CORE_FORMS_PERMISSION.codename],
-            "userRoles": [],
-            "userRolesPermissions": [],
-        }
-
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": new_profile_id, "version": "v2"}), data=new_data, format="json"
-        )
-        self.assertJSONResponse(response, 200)
-
-        response = self.client.get(
-            reverse("logs-list"),
-            {
-                "contentType": "iaso.profile",
-                "fields": ",".join(["past_value", "new_value"]),
-                "objectId": new_profile_id,
-            },
-        )
-        response_data = self.assertJSONResponse(response, 200)
-        logs = response_data["list"]
-        log = logs[0]
-
-        try:
-            jsonschema.validate(instance=log, schema=PROFILE_LOG_SCHEMA)
-        except jsonschema.exceptions.ValidationError as ex:
-            self.fail(msg=str(ex))
-        self.assertEqual(log["past_value"][0]["pk"], new_profile_id)
-        past_value = log["past_value"][0]["fields"]
-        self.assertEqual(past_value["user"], new_user_id)
-        self.assertEqual(past_value["username"], new_user_name)
-        self.assertEqual(past_value["first_name"], data["firstName"])
-        self.assertEqual(past_value["last_name"], data["lastName"])
-        self.assertEqual(past_value["email"], data["email"])
-        self.assertEqual(len(past_value["user_permissions"]), 1)
-        self.assertEqual(past_value["user_permissions"], data["userPermissions"])
-        self.assertNotIn("password", past_value.keys())
-
-        self.assertEqual(past_value["dhis2_id"], data["dhis2Id"])
-        self.assertEqual(past_value["language"], data["language"])
-        self.assertEqual(past_value["home_page"], data["homePage"])
-        self.assertEqual(past_value["phone_number"], data["phoneNumber"])
-        self.assertEqual(len(past_value["org_units"]), 1)
-        self.assertIn(
-            self.org_unit_from_parent_type.id,
-            past_value["org_units"],
-        )
-        self.assertEqual(len(past_value["user_roles"]), 1)
-        self.assertIn(self.user_role.id, past_value["user_roles"])
-        self.assertEqual(len(past_value["projects"]), 1)
-        self.assertIn(self.project.id, past_value["projects"])
-        self.assertEqual(log["new_value"][0]["pk"], new_profile_id)
-        new_value = log["new_value"][0]["fields"]
-        self.assertEqual(len(new_value["user_permissions"]), 2)
-        self.assertIn("iaso_forms", new_value["user_permissions"])
-        self.assertIn("iaso_org_units_read", new_value["user_permissions"])
-        self.assertEqual(new_value["language"], new_data["language"])
-        self.assertEqual(new_value["home_page"], new_data["homePage"])
-        self.assertEqual(new_value["organization"], new_data["organization"])
-        self.assertEqual(len(new_value["org_units"]), 2)
-        self.assertIn(self.org_unit_from_parent_type.id, new_value["org_units"])
-        self.assertIn(self.org_unit_from_sub_type.id, new_value["org_units"])
-        self.assertEqual(new_value["user_roles"], [])
-
-    def test_log_on_user_updates_own_profile(self):
-        self.client.force_authenticate(self.jim)
-        new_data = {"language": "fr"}
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": PK_ME, "version": "v2"}), data=new_data, format="json"
-        )
-        self.assertJSONResponse(response, 200)
-        # Log as super user to access the logs API
-        self.client.force_authenticate(self.john)
-        response = self.client.get(
-            reverse("logs-list"),
-            data={
-                "contentType": "iaso.profile",
-                "fields": ",".join(["past_value", "new_value"]),
-                "objectId": self.jim.iaso_profile.id,
-            },
-        )
-        response_data = self.assertJSONResponse(response, 200)
-        logs = response_data["list"]
-        log = logs[0]
-
-        try:
-            jsonschema.validate(instance=log, schema=PROFILE_LOG_SCHEMA)
-        except jsonschema.exceptions.ValidationError as ex:
-            self.fail(msg=str(ex))
-
-        self.assertEqual(log["past_value"][0]["pk"], self.jim.iaso_profile.id)
-        past_value = log["past_value"][0]["fields"]
-        self.assertEqual(past_value["user"], self.jim.id)
-        self.assertEqual(past_value["username"], self.jim.username)
-        self.assertEqual(past_value["first_name"], "")
-        self.assertEqual(past_value["language"], None)
-        self.assertNotIn("password", past_value.keys())
-
-        self.assertEqual(log["new_value"][0]["pk"], self.jim.iaso_profile.id)
-        new_value = log["new_value"][0]["fields"]
-        self.assertEqual(new_value["user"], self.jim.id)
-        self.assertEqual(new_value["username"], self.jim.username)
-        self.assertEqual(new_value["first_name"], "")
-        self.assertEqual(new_value["language"], "fr")
-        self.assertNotIn("password", new_value.keys())
-
-    def test_profile_list_search_by_children_ou_deep_hierarchy(self):
-        """Test that searching by children org units returns profiles from all levels of the hierarchy"""
-        # Create a deeper hierarchy
-        child_of_child = m.OrgUnit.objects.create(
-            org_unit_type=self.parent_org_unit_type,
-            version=self.account.default_version,
-            name="Child of child org unit",
-            parent=self.child_org_unit,
-        )
-
-        grand_child = m.OrgUnit.objects.create(
-            org_unit_type=self.parent_org_unit_type,
-            version=self.account.default_version,
-            name="Grand child org unit",
-            parent=child_of_child,
-        )
-
-        # Clear existing org units first to avoid interference
-        for profile in m.Profile.objects.all():
-            profile.org_units.clear()
-
-        # Assign users to different levels of the hierarchy
-        self.jane.iaso_profile.org_units.set([self.org_unit_from_parent_type])  # Root
-        self.jim.iaso_profile.org_units.set([self.child_org_unit])  # Child
-        self.jam.iaso_profile.org_units.set([child_of_child])  # Child of child
-        self.jom.iaso_profile.org_units.set([grand_child])  # Grand child
-
-        self.client.force_authenticate(self.jane)
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v2"}),
-            {"location": self.org_unit_from_parent_type.pk, "ouChildren": True},
-        )
-
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 4)
-        profiles = response_data["results"]
-        usernames = [p["userName"] for p in profiles]
-
-        # Should include users from all levels
-        self.assertIn("janedoe", usernames)  # Root level
-        self.assertIn("jim", usernames)  # Child level
-        self.assertIn("jam", usernames)  # Child of child level
-        self.assertIn("jom", usernames)  # Grand child level
-
-    def test_profile_list_search_by_parent_and_children_ou(self):
-        """Test that searching with both parent and children flags returns the complete hierarchy"""
-        # Clear existing org units first
-        for profile in m.Profile.objects.all():
-            profile.org_units.clear()
-
-        # Setup hierarchy
-        parent_of_root = m.OrgUnit.objects.create(
-            org_unit_type=self.parent_org_unit_type,
-            version=self.account.default_version,
-            name="Parent of root",
-        )
-        self.org_unit_from_parent_type.parent = parent_of_root
-        self.org_unit_from_parent_type.save()
-
-        # Assign users to different levels
-        self.jane.iaso_profile.org_units.set([parent_of_root])  # Parent
-        self.jim.iaso_profile.org_units.set([self.org_unit_from_parent_type])  # Current
-        self.jam.iaso_profile.org_units.set([self.child_org_unit])  # Child
-
-        self.client.force_authenticate(self.jane)
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v2"}),
-            {"location": self.org_unit_from_parent_type.pk, "ouParent": True, "ouChildren": True},
-        )
-
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 3)
-        profiles = response_data["results"]
-        usernames = [p["userName"] for p in profiles]
-
-        # Should include users from parent, current and child levels
-        self.assertIn("janedoe", usernames)  # Parent level
-        self.assertIn("jim", usernames)  # Current level
-        self.assertIn("jam", usernames)  # Child level
-
-    def test_profile_list_search_by_children_ou_no_duplicates(self):
-        """Test that searching by children org units doesn't return duplicate profiles
-        when a user is assigned to multiple levels"""
-        # Clear existing org units first
-        for profile in m.Profile.objects.all():
-            profile.org_units.clear()
-
-        # Assign a user to multiple levels in the hierarchy
-        self.jane.iaso_profile.org_units.set(
-            [
-                self.org_unit_from_parent_type,  # Root
-                self.child_org_unit,  # Child
-            ]
-        )
-
-        self.client.force_authenticate(self.jane)
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v2"}),
-            {"location": self.org_unit_from_parent_type.pk, "ouChildren": True},
-        )
-
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 1)
-
-        # Should only include the user once despite being in multiple levels
-        self.assertEqual(response_data["results"][0]["userName"], "janedoe")
-
-    def test_profile_list_search_with_children_ou_preserves_search_results(self):
-        """
-        Test that search results are preserved with `ouChildren`.
-        """
-        # Clear existing org units first.
-        for profile in m.Profile.objects.all():
-            profile.org_units.clear()
-
-        # This user matches "jim" search.
-        self.jim.iaso_profile.org_units.set([self.child_org_unit])
-
-        # Create a user that matches "jim" search but is NOT in the hierarchy.
-        self.create_user_with_profile(
-            username="jim_outside", account=self.account, permissions=[CORE_USERS_ADMIN_PERMISSION]
-        )
-
-        self.client.force_authenticate(self.jim)
-
-        # Search for "jim" without ouChildren - should return both `jim` and `jim_outside`.
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}), {"search": "jim"})
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 2)
-        profiles = response_data["results"]
-        self.assertEqual(profiles[0]["userName"], "jim")
-        self.assertEqual(profiles[1]["userName"], "jim_outside")
-
-        # Search for "jim" with `ouChildren=true` - should only return jim (who is in hierarchy).
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v2"}),
-            {"search": "jim", "location": self.org_unit_from_parent_type.pk, "ouChildren": True},
-        )
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 1)
-        profiles = response_data["results"]
-        self.assertEqual(profiles[0]["userName"], "jim")
-
-    def test_update_password_for_single_user(self):
-        single_user = self.jim
-        single_user.set_password("p4ssword")
-        single_user.save()
-
-        self.client.force_authenticate(single_user)
-        new_data = {"password": "new_p4ssword", "confirmPassword": "new_p4ssword"}
-        response = self.client.patch(
-            reverse("profiles-update-password", kwargs={"pk": single_user.iaso_profile.pk, "version": "v2"}),
-            data=new_data,
-            format="json",
-        )
-        self.assertJSONResponse(response, 204)
-        single_user.refresh_from_db()
-        self.assertEqual(single_user.check_password("new_p4ssword"), True)
-
-    def test_update_password_for_multi_user(self):
-        """
-        For tenant users, changing the password of an `account_user`
-        should update the password of the `main_user`.
-        """
-        main_user = m.User.objects.create(username="main_user", email="main_user@health.org")
-        main_user.set_password("p4ssword")
-        main_user.save()
-        account_user = self.create_user_with_profile(
-            username="user_1",
-            email="user_1@health.org",
-            account=self.account,
-            permissions=[CORE_USERS_ADMIN_PERMISSION],
-        )
-        m.TenantUser.objects.create(main_user=main_user, account_user=account_user)
-
-        self.client.force_authenticate(account_user)
-        new_data = {"password": "new_p4ssword", "confirmPassword": "new_p4ssword"}
-        response = self.client.patch(
-            reverse("profiles-update-password", kwargs={"pk": account_user.iaso_profile.pk, "version": "v2"}),
-            data=new_data,
-            format="json",
-        )
-        self.assertJSONResponse(response, 204)
-        main_user.refresh_from_db()
-        self.assertEqual(main_user.check_password("new_p4ssword"), True)
-
-    def test_list_profiles_sorted_by_annotated_first_user_role(self):
-        """
-        Test that profiles are properly sorted by their alphabetically first user role.
-        """
-        data_manager_group = Group.objects.create(name=f"{self.account.pk}_Data manager")
-        gpei_group = Group.objects.create(name=f"{self.account.pk}_GPEI coordinators")
-        zulu_group = Group.objects.create(name=f"{self.account.pk}_Zulu role")
-
-        data_manager_role = m.UserRole.objects.create(group=data_manager_group, account=self.account)
-        gpei_role = m.UserRole.objects.create(group=gpei_group, account=self.account)
-        zulu_role = m.UserRole.objects.create(group=zulu_group, account=self.account)
-
-        # Clear the users.
-        self.account.profile_set.exclude(user_id=self.john.pk).delete()
-
-        # User 1: multiple roles starting with "Data manager" (should be first).
-        user1 = self.create_user_with_profile(
-            username="user_multi_data", account=self.account, permissions=[CORE_USERS_ADMIN_PERMISSION]
-        )
-        user1.iaso_profile.user_roles.set([data_manager_role, gpei_role])
-
-        # User 2: single role "GPEI" (should be middle).
-        user2 = self.create_user_with_profile(
-            username="user_single_gpei", account=self.account, permissions=[CORE_USERS_ADMIN_PERMISSION]
-        )
-        user2.iaso_profile.user_roles.set([gpei_role])
-
-        # User 3: single role "Zulu" (should be last)
-        user3 = self.create_user_with_profile(
-            username="user_single_zulu", account=self.account, permissions=[CORE_USERS_ADMIN_PERMISSION]
-        )
-        user3.iaso_profile.user_roles.set([zulu_role])
-
-        # User 4: multiple roles starting with "GPEI coordinators" (should be between user1 and user3)
-        user4 = self.create_user_with_profile(
-            username="user_multi_gpei", account=self.account, permissions=[CORE_USERS_ADMIN_PERMISSION]
-        )
-        user4.iaso_profile.user_roles.set([gpei_role, zulu_role])
-
-        self.client.force_authenticate(self.john)
-
-        # Test ascending order.
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v2"}), {"order": "annotatedFirstUserRole", "limit": 10}
-        )
-        response_data = self.assertJSONResponse(response, 200)
-
-        actual_order = []
-        for profile in response_data["results"]:
-            if not profile["userRoles"]:
-                actual_order.append(None)
-                continue
-            role_id = profile["userRoles"][0]
-            group_name = m.UserRole.objects.get(id=role_id["id"]).group.name
-            actual_order.append(group_name)
-
-        expected_order = [
-            f"{self.account.pk}_Data manager",
-            f"{self.account.pk}_GPEI coordinators",
-            f"{self.account.pk}_GPEI coordinators",
-            f"{self.account.pk}_Zulu role",
-            None,
-        ]
-
-        self.assertEqual(
-            actual_order,
-            expected_order,
-            f"Users not sorted correctly by first role. Expected: {expected_order}, got: {actual_order}",
-        )
-
-        # Test descending order.
-
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v2"}), {"order": "-annotatedFirstUserRole", "limit": 10}
-        )
-        response_data = self.assertJSONResponse(response, 200)
-
-        actual_order = []
-        for profile in response_data["results"]:
-            if not profile["userRoles"]:
-                actual_order.append(None)
-                continue
-            role_id = profile["userRoles"][0]
-            group_name = m.UserRole.objects.get(id=role_id["id"]).group.name
-            actual_order.append(group_name)
-
-        expected_order = [
-            None,
-            f"{self.account.pk}_Zulu role",
-            f"{self.account.pk}_GPEI coordinators",
-            f"{self.account.pk}_GPEI coordinators",
-            f"{self.account.pk}_Data manager",
-        ]
-
-        self.assertEqual(
-            actual_order,
-            expected_order,
-            f"Users not sorted correctly by first role. Expected: {expected_order}, got: {actual_order}",
-        )
-
-    def test_update_profile_without_changing_username_should_succeed(self):
-        """Test that updating profile without changing username succeeds."""
-        self.client.force_authenticate(self.john)
-
-        alice = self.create_user_with_profile(
-            username="alice", account=self.account, first_name="Alice", last_name="Smith"
-        )
-        bob = self.create_user_with_profile(username="bob", account=self.account, first_name="Bob", last_name="Smith")
-
-        alice_profile = Profile.objects.get(user=alice)
-        data = {
-            "userName": alice.username,
-            "firstName": "Alice Changed",
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": alice_profile.id, "version": "v2"}), data=data, format="json"
-        )
-
-        self.assertJSONResponse(response, 200)
-        alice.refresh_from_db()
-        self.assertEqual(alice.username, "alice")
-        self.assertEqual(alice.first_name, "Alice Changed")
-
-        bob_profile = Profile.objects.get(user=bob)
-        data = {
-            "firstName": "Bob Changed",
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": bob_profile.id, "version": "v2"}), data=data, format="json"
-        )
-
-        self.assertJSONResponse(response, 200)
-        bob.refresh_from_db()
-        self.assertEqual(bob.username, "bob")
-        self.assertEqual(bob.first_name, "Bob Changed")
-
-    def test_update_profile_to_existing_username_should_fail(self):
-        """Test that changing username to an existing one (case-insensitive) fails correctly."""
-        self.client.force_authenticate(self.john)
-
-        alice = self.create_user_with_profile(
-            username="alice", account=self.account, first_name="Alice", last_name="Smith"
-        )
-        bob = self.create_user_with_profile(username="bob", account=self.account, first_name="Bob", last_name="Wilson")
-
-        # Try to change bob's username to "Alice"
-        bob_profile = Profile.objects.get(user=bob)
-        data = {
-            "userName": "Alice",
-            "firstName": "Bob Updated",
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": bob_profile.id, "version": "v2"}), data=data, format="json"
-        )
-
-        response_data = self.assertJSONResponse(response, 400)
-        self.assertHasError(response_data, "userName", "Username already exists")
-
-        bob.refresh_from_db()
-        self.assertEqual(bob.username, "bob")
-        self.assertEqual(bob.first_name, "Bob")
-
-    def test_update_profile_with_case_insensitive_username_should_succeed(self):
-        """Test that updating profile without changing username succeeds even with case-insensitive matching usernames."""
-        self.client.force_authenticate(self.john)
-
-        alice_lower = self.create_user_with_profile(
-            username="alice", account=self.account, first_name="Alice Lower", last_name="alice lower"
-        )
-        alice_upper = self.create_user_with_profile(
-            username="Alice", account=self.account, first_name="Alice Upper", last_name="alice upper"
-        )
-
-        alice_lower_profile = Profile.objects.get(user=alice_lower)
-        data = {
-            "userName": alice_lower.username,
-            "firstName": "Alice Lower Changed",
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": alice_lower_profile.id, "version": "v2"}), data=data, format="json"
-        )
-
-        self.assertJSONResponse(response, 200)
-        alice_lower.refresh_from_db()
-        self.assertEqual(alice_lower.username, "alice")
-        self.assertEqual(alice_lower.first_name, "Alice Lower Changed")
-
-        alice_upper_profile = Profile.objects.get(user=alice_upper)
-        data = {
-            "userName": alice_upper.username,
-            "firstName": "Alice Upper Changed",
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": alice_upper_profile.id, "version": "v2"}), data=data, format="json"
-        )
-
-        self.assertJSONResponse(response, 200)
-        alice_upper.refresh_from_db()
-        self.assertEqual(alice_upper.username, "Alice")
-        self.assertEqual(alice_upper.first_name, "Alice Upper Changed")
-
-    def test_update_username_to_existing_case_variation_should_fail(self):
-        """Test that changing username to case variation of another user's username fails."""
-        self.client.force_authenticate(self.john)
-
-        alice_lower = self.create_user_with_profile(
-            username="alice", account=self.account, first_name="Alice Lower", last_name="Smith"
-        )
-        alice_upper = self.create_user_with_profile(
-            username="Alice", account=self.account, first_name="Alice Upper", last_name="Jones"
-        )
-
-        alice_lower_profile = Profile.objects.get(user=alice_lower)
-        data = {
-            "userName": "Alice",
-            "firstName": "Alice Changed",
-        }
-        response = self.client.patch(
-            reverse("profiles-detail", kwargs={"pk": alice_lower_profile.id, "version": "v2"}), data=data, format="json"
-        )
-
-        # Should fail because another user already has "Alice"
-        response_data = self.assertJSONResponse(response, 400)
-        self.assertHasError(response_data, "userName", "Username already exists")
-
-        # Ensure original username wasn't changed
-        alice_lower.refresh_from_db()
-        self.assertEqual(alice_lower.username, "alice")
-        self.assertEqual(alice_lower.first_name, "Alice Lower")
-
-        # Ensure other username wasn't changed
-        alice_upper.refresh_from_db()
-        self.assertEqual(alice_upper.username, "Alice")
-        self.assertEqual(alice_upper.first_name, "Alice Upper")
-
-    def test_create_user_with_invitation_sets_random_password(self):
-        """
-        Test that creating a user with send_email_invitation=True sets a random password instead of an unusable one.
-        """
-        self.client.force_authenticate(self.jim)
-        data = {
-            "userName": "invited_user_empty_password",
-            "password": "",
-            "firstName": "Invited",
-            "lastName": "User",
-            "sendEmailInvitation": True,
-            "email": "invited1@test.com",
-        }
-
-        with self.captureOnCommitCallbacks(execute=True) as callbacks:
-            response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        self.assertEqual(len(callbacks), 1)
-        self.assertJSONResponse(response, 201)
-
-        user = get_user_model().objects.get(username="invited_user_empty_password")
-        self.assertTrue(user.has_usable_password(), "Invited user should have a usable password")
-
-        data = {
-            "userName": "invited_user_missing_password",
-            "firstName": "Invited",
-            "lastName": "User",
-            "sendEmailInvitation": True,
-            "email": "invited2@test.com",
-        }
-
-        with self.captureOnCommitCallbacks(execute=True) as callbacks:
-            response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-        self.assertEqual(len(callbacks), 1)
-        self.assertJSONResponse(response, 201)
-
-        user = get_user_model().objects.get(username="invited_user_missing_password")
-        self.assertTrue(user.has_usable_password())
-
-        data = {
-            "userName": "invited_user_password_is_none",
-            "password": None,
-            "firstName": "Invited",
-            "lastName": "User",
-            "sendEmailInvitation": True,
-            "email": "invited3@test.com",
-        }
-
-        with self.captureOnCommitCallbacks(execute=True) as callbacks:
-            response = self.client.post(reverse("profiles-list", kwargs={"version": "v2"}), data=data, format="json")
-
-        self.assertJSONResponse(response, 201)
-        self.assertEqual(len(callbacks), 1)
-
-        user = get_user_model().objects.get(username="invited_user_password_is_none")
-        self.assertTrue(user.has_usable_password())
-
-        self.assertEqual(len(mail.outbox), 3)
-        self.assertEqual(mail.outbox[0].to, ["invited1@test.com"])
-        self.assertEqual(mail.outbox[1].to, ["invited2@test.com"])
-        self.assertEqual(mail.outbox[2].to, ["invited3@test.com"])
-
-    def test_search_parameters_v1(self):
-        self.client.force_authenticate(self.jane)
-        self.jane.iaso_profile.org_units.set([self.org_unit_from_parent_type])
-
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v1"}), {"limit": 100})
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertEqual(response_data["count"], 7)
-
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v1"}),
-            {"org_unit_types": self.parent_org_unit_type.pk, "limit": 100},
-        )
-
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertEqual(response_data["count"], 2)
-        self.assertEqual(response_data["results"][0]["user_name"], "janedoe")
-
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v1"}),
-            {"orgUnitTypes": self.parent_org_unit_type.pk, "limit": 100},
-        )
-
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertEqual(response_data["count"], 7)
-        self.assertEqual(response_data["results"][0]["user_name"], "janedoe")
-
-    def test_search_parameters_default(self):
-        self.client.force_authenticate(self.jane)
-        self.jane.iaso_profile.org_units.set([self.org_unit_from_parent_type])
-
-        response = self.client.get(reverse("profiles-list"), {"limit": 100})
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertEqual(response_data["count"], 7)
-
-        response = self.client.get(
-            reverse("profiles-list"),
-            {"org_unit_types": self.parent_org_unit_type.pk, "limit": 100},
-        )
-
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertEqual(response_data["count"], 2)
-        self.assertEqual(response_data["results"][0]["user_name"], "janedoe")
-
-        response = self.client.get(
-            reverse("profiles-list"),
-            {"orgUnitTypes": self.parent_org_unit_type.pk, "limit": 100},
-        )
-
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertEqual(response_data["count"], 7)
-        self.assertEqual(response_data["results"][0]["user_name"], "janedoe")
-
-    def test_search_users_by_organization(self):
-        self.jane.iaso_profile.organization = "Some organization"
-        self.jane.iaso_profile.save()
-
-        self.client.force_authenticate(self.jane)
-        response = self.client.get(reverse("profiles-list", kwargs={"version": "v2"}), {"limit": 100})
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 7)
-
-        for parameter in ["so", "some org", "Some organization"]:
-            with self.subTest(f"Searching with {parameter}"):
-                response = self.client.get(
-                    reverse("profiles-list", kwargs={"version": "v2"}), {"limit": 100, "search": parameter}
-                )
-                response_data = self.assertJSONResponse(response, 200)
-                self.assertValidProfileListData(response_data, 1)
-
-        response = self.client.get(
-            reverse("profiles-list", kwargs={"version": "v2"}), {"limit": 100, "search": "wrong search"}
-        )
-        response_data = self.assertJSONResponse(response, 200)
-        self.assertValidProfileListData(response_data, 0)
-
-    def test_dropdown_endpoint(self):
-        self.jane.iaso_profile.organization = "Some organization"
-        self.jane.iaso_profile.save()
-        self.client.force_authenticate(self.jane)
-
-        with self.subTest("should not be paginated by default"):
-            response = self.client.get(reverse("profiles-dropdown"))
-
-            response_data = self.assertJSONResponse(response, 200)
-            self.assertValidListData(list_data=response_data, results_key=None, expected_length=7, paginated=False)
-
-        with self.subTest("should not be paginated if limit is provided"):
-            response = self.client.get(reverse("profiles-dropdown"), {"limit": 1})
-            response_data = self.assertJSONResponse(response, 200)
-            self.assertValidListData(list_data=response_data, results_key="results", expected_length=1, paginated=True)
-
-        with self.subTest("Can still be filtered"):
-            for parameter in ["so", "some org", "Some organization"]:
-                with self.subTest(f"Searching with {parameter}"):
-                    response = self.client.get(reverse("profiles-dropdown"), {"search": parameter})
-                    response_data = self.assertJSONResponse(response, 200)
-                    self.assertValidListData(
-                        list_data=response_data, results_key=None, expected_length=1, paginated=False
-                    )
-
-        with self.subTest("Check response data"):
-            response = self.client.get(reverse("profiles-dropdown"), {"search": parameter})
-            response_data = self.assertJSONResponse(response, 200)
-            self.assertValidListData(list_data=response_data, results_key=None, expected_length=1, paginated=False)
-
-            item = response_data[0]
-            self.assertCountEqual(item.keys(), ["label", "value"])
-            self.assertEqual(item["value"], self.jane.id)
-            self.assertEqual(item["label"], f"{self.jane.username} ({self.jane.get_full_name()})")
